% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalMetrics.R
\name{barErrorPlot}
\alias{barErrorPlot}
\title{Generate bar error plot and its dispersion by cell types or by number of
different cell types in test bulk samples.}
\usage{
barErrorPlot(
  object,
  error,
  by,
  dispersion = "se",
  filter.sc = TRUE,
  title = NULL,
  angle = 90,
  theme = theme_grey()
)
}
\arguments{
\item{object}{\code{DigitalDLSorter} object with \code{trained.model} slot
containing metrics in \code{eval.stats.samples} slot.}

\item{error}{MAE or MSE. By default it is used a list of custom colors
provided by the package.}

\item{by}{Variable used to display errors.}

\item{dispersion}{Standard error (\code{'se'}) or standard deviation
(\code{'sd'}). The first by default.}

\item{filter.sc}{Boolean indicating if filter single-cell profiles and only
display correlations of results associated with bulk samples (\code{TRUE}
by default).}

\item{title}{Title of the plot.}

\item{angle}{Angle of ticks.}

\item{theme}{ggplot theme.}
}
\description{
Generate bar error plot and its dispersion by cell types (\code{CellType}) or
by number of different cell types (\code{nMix}) in test bulk samples.
}
\examples{
barErrorPlot(
  object = DDLSSmallCompleted,
  error = "MSE",
  by = "CellType"
)

barErrorPlot(
  object = DDLSSmallCompleted,
  error = "MAE",
  by = "nMix"
)
}
\seealso{
\code{\link{calculateEvalMetrics}} \code{\link{corrExpPredPlot}}
\code{\link{distErrorPlot}} \code{\link{blandAltmanLehPlot}}
}
