% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalMetrics.R
\name{blandAltmanLehPlot}
\alias{blandAltmanLehPlot}
\title{Generate Bland-Altman agreement plots between predicted and expected cell
type proportions from test data results}
\usage{
blandAltmanLehPlot(
  object,
  colors,
  color.by = "CellType",
  facet.by = NULL,
  log.2 = FALSE,
  filter.sc = TRUE,
  density = TRUE,
  color.density = "darkblue",
  size.point = 0.05,
  alpha.point = 1,
  ncol = NULL,
  nrow = NULL,
  title = NULL,
  theme = NULL,
  ...
)
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object with
\code{trained.model} slot containing metrics in \code{test.deconv.metrics}
slot.}

\item{colors}{Vector of colors to be used. Only vectors with a number of
colors equal to or greater than the levels of \code{color.by} will be
accepted. By default a custom color list is used.}

\item{color.by}{Variable used to color data. Options are \code{nCellTypes}
and \code{CellType}.}

\item{facet.by}{Variable used to display the data in different panels. If
\code{NULL}, the plot is not split into different panels. Options are
\code{nCellTypes} (by number of different cell types) and \code{CellType}
(by cell type).}

\item{log.2}{Whether to display the Bland-Altman agreement plot in log2 space
(\code{FALSE} by default).}

\item{filter.sc}{Boolean indicating whether single-cell profiles are filtered
out and only correlations of results associated with bulk samples are
displayed (\code{TRUE} by default).}

\item{density}{Boolean indicating whether density lines must be displayed
(\code{TRUE} by default).}

\item{color.density}{Color of density lines if the \code{density} argument is
\code{TRUE}.}

\item{size.point}{Size of the points (0.1 by default).}

\item{alpha.point}{Alpha of the points (0.1 by default).}

\item{ncol}{Number of columns if \code{facet.by} is used.}

\item{nrow}{Number of rows if \code{facet.by} is used.}

\item{title}{Title of the plot.}

\item{theme}{\pkg{ggplot2} theme.}

\item{...}{Additional argument for the \code{facet_wrap} function from
\pkg{ggplot2} if \code{facet.by} is not \code{NULL}.}
}
\value{
A ggplot object with Bland-Altman agreement plots between expected
and actual proportions.
}
\description{
Generate Bland-Altman agreement plots between predicted and expected cell
type proportions from test data results. The Bland-Altman agreement plots can
be displayed all mixed or split by cell type (\code{CellType}) or the number
of cell types present in samples (\code{nCellTypes}). See the \code{facet.by}
argument and examples for more information.
}
\examples{
\dontrun{
set.seed(123)
sce <- SingleCellExperiment::SingleCellExperiment(
  assays = list(
    counts = matrix(
      rpois(30, lambda = 5), nrow = 15, ncol = 20,
      dimnames = list(paste0("Gene", seq(15)), paste0("RHC", seq(20)))
    )
  ),
  colData = data.frame(
    Cell_ID = paste0("RHC", seq(20)),
    Cell_Type = sample(x = paste0("CellType", seq(6)), size = 20,
                       replace = TRUE)
  ),
  rowData = data.frame(
    Gene_ID = paste0("Gene", seq(15))
  )
)
DDLS <- loadSCProfiles(
  single.cell.data = sce,
  cell.ID.column = "Cell_ID",
  gene.ID.column = "Gene_ID"
)
probMatrixValid <- data.frame(
  Cell_Type = paste0("CellType", seq(6)),
  from = c(1, 1, 1, 15, 15, 30),
  to = c(15, 15, 30, 50, 50, 70)
)
DDLS <- generateBulkCellMatrix(
  object = DDLS,
  cell.ID.column = "Cell_ID",
  cell.type.column = "Cell_Type",
  prob.design = probMatrixValid,
  num.bulk.samples = 50,
  verbose = TRUE
)
# training of DDLS model
tensorflow::tf$compat$v1$disable_eager_execution()
DDLS <- trainDigitalDLSorterModel(
  object = DDLS,
  on.the.fly = TRUE,
  batch.size = 15,
  num.epochs = 5
)
# evaluation using test data
DDLS <- calculateEvalMetrics(
  object = DDLS
)
# Bland-Altman plot by cell type
blandAltmanLehPlot(
  object = DDLS,
  facet.by = "CellType",
  color.by = "CellType"
)
# Bland-Altman plot of all samples mixed
blandAltmanLehPlot(
  object = DDLS,
  facet.by = NULL,
  color.by = "CellType",
  alpha.point = 0.3,
  log2 = TRUE
)
}

}
\seealso{
\code{\link{calculateEvalMetrics}} \code{\link{corrExpPredPlot}}
\code{\link{distErrorPlot}} \code{\link{barErrorPlot}}
}
