% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{bulk.simul}
\alias{bulk.simul}
\alias{bulk.simul<-}
\title{Get and set \code{bulk.simul} slot in a \code{DigitalDLSorter}
object.}
\usage{
bulk.simul(object, type.data = "both")

bulk.simul(object, type.data = "both") <- value
}
\arguments{
\item{object}{A \code{DigitalDLSorter} object.}

\item{type.data}{Element of the list. Can be 'train', 'test' or 'both' (the
last by default).}

\item{value}{A \code{list} with two elements, train and test, each one being
a \code{SummarizedExperiment} object with simulated bulk RNA-Seq samples.}
}
\description{
Get and set \code{bulk.simul} slot in a \code{DigitalDLSorter}
object.
}
