% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{cell.names}
\alias{cell.names}
\alias{cell.names,ProbMatrixCellTypes-method}
\alias{cell.names<-}
\alias{cell.names<-,ProbMatrixCellTypes-method}
\title{Get and set \code{cell.names} slot in a
\code{\linkS4class{ProbMatrixCellTypes}} object}
\usage{
cell.names(object)

cell.names(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{ProbMatrixCellTypes}} object.}

\item{value}{Matrix containing bulk samples as rows and cells that
will be used for simulating these samples as columns (\code{n.cell}
argument)}
}
\description{
Get and set \code{cell.names} slot in a
\code{\linkS4class{ProbMatrixCellTypes}} object
}
