% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{cell.types}
\alias{cell.types}
\alias{cell.types,DigitalDLSorterDNN-method}
\alias{cell.types<-}
\alias{cell.types<-,DigitalDLSorterDNN-method}
\title{Get and set \code{cell.types} slot in a
\code{\linkS4class{DigitalDLSorterDNN}} object}
\usage{
cell.types(object)

cell.types(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorterDNN}} object.}

\item{value}{Vector with cell types considered by the Deep Neural Network
model.}
}
\description{
Get and set \code{cell.types} slot in a
\code{\linkS4class{DigitalDLSorterDNN}} object
}
