% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colorectal.li}
\alias{colorectal.li}
\title{Pre-trained \code{\linkS4class{DigitalDLSorter}} DNN model for deconvolution
of tumor-infiltrating lymphocytes (TILs) present in colorectal cancer
environment}
\format{
A \code{\linkS4class{DigitalDLSorterDNN}} object with the following
slots: \describe{ \item{model}{Trained DNN model.}
\item{training.history}{Evolution of metrics and loss function during
training.} \item{test.metrics}{Metrics and loss results on test data.}
\item{test.pred}{Predictions on test data.} \item{cell.types}{Cell types
considered by the model.} \item{features}{Features (genes) considered by
the model.} }
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE81861}
}
\usage{
colorectal.li
}
\description{
\code{\linkS4class{DigitalDLSorter}} DNN model built and trained with
single-cell data from Li et al., 2017 (GSE81861). This model allows the
enumeration and quantification of immune infiltrated cell types in colorectal
cancer environment. This data set consists of single-cell RNA-seq profiles
from 11 patients from different tumor etiology and stages (see Torroja and
Sanchez-Cabo, 2019 and Li et al., 2017 for more details). The analysis and
characterization of cells was carried out by the authors of
\pkg{digitalDLSorteR}.
}
\details{
This model considers 10 cell types: cancer cells (CRC), epithelial (Ep),
monocytes (M), macrophages (Mc), fibroblast (Fb), CD4 T cells (CD4), CD8 T
cells (CD8Gp), CD8 T cells (CD8Gn), germinal center B cells (gB) and mature B
cells (pB).

The genes considered are 36477 in SYMBOL notation.

The model consists of 2 hidden layers with 200 neurons per layer trained with
'kullback_leibler_divergence' loss function, batch size equal to 128 and a
number of epochs equal to 30.
}
\references{
Li, H., Courtois, E. T., Sengupta, D., Tan, Y., Chen, K. H., Goh,
J. J. L., et al. (2017). Reference component analysis of single-cell
transcriptomes elucidates cellular heterogeneity in human colorectal
tumors. Nat. Genet. 49 (5), 708–718. doi: \url{10.1038/ng.3818}

Torroja, C. y Sánchez-Cabo, F. (2019). digitalDLSorter: A Deep Learning
algorithm to quantify immune cell populations based on scRNA-Seq data.
Frontiers in Genetics 10, 978. doi: \url{10.3389/fgene.2019.00978}
}
\keyword{datasets}
