% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalMetrics.R
\name{corrExpPredPlot}
\alias{corrExpPredPlot}
\title{Generate correlation plot between predicted and expected cell type
proportions from test samples.}
\usage{
corrExpPredPlot(
  object,
  colors,
  facet.by = NULL,
  color.by = "CellType",
  corr = "both",
  filter.sc = TRUE,
  pos.x.label = 0.01,
  pos.y.label = 0.95,
  sep.labels = 0.15,
  size.point = 0.1,
  alpha.point = 1,
  ncol = NULL,
  nrow = NULL,
  title = NULL,
  theme = theme_grey(),
  ...
)
}
\arguments{
\item{object}{\code{DigitalDLSorter} object with \code{trained.model} slot
containing metrics in \code{eval.stats.samples} slot.}

\item{colors}{Vector of colors to use. Only vectors with a number of colors
equal to or greater than the levels of \code{color.by} will be accepted. By
default it is used a list of custom colors provided by the package.}

\item{facet.by}{Variable used to display data in different panels. If it is
\code{NULL}, the plot is not separated into different panels. The options
are \code{nMix} (by number of different cell types) and \code{CellType} (by
cell type).}

\item{color.by}{Variable used to color data. The options are \code{nMix} and
\code{CellType}.}

\item{corr}{Correlation value displayed as annotation. The available metrics
are Pearson's correlation coefficient (\code{'pearson'}) and concordance
correlation coefficient (\code{'ccc'}). The argument can be equal to
\code{'pearson'}, \code{'ccc'} or \code{'both'} (by default).}

\item{filter.sc}{Boolean indicating if filter single-cell profiles and only
display correlations of results associated with bulk samples (\code{TRUE}
by default).}

\item{pos.x.label}{Position on the X axis of the errors annotations. 0.95 by
default.}

\item{pos.y.label}{Position on the Y axis of the errors annotations. 0.1 by
default.}

\item{sep.labels}{Space separating annotations if \code{corr} is equal to
\code{'both'}. 0.15 by default.}

\item{size.point}{Size of points (0.1 by default).}

\item{alpha.point}{Alpha of points (0.1 by default).}

\item{ncol}{Number of columns if \code{facet.by} is different than
\code{NULL}.}

\item{nrow}{Number of rows if \code{facet.by} is different than \code{NULL}.}

\item{title}{Title of the plot.}

\item{theme}{ggplot theme.}

\item{...}{Additional argument for \code{facet_wrap} ggplot function if
\code{facet.by} is not equal to \code{NULL}.}
}
\description{
Generate correlation plot between predicted and expected cell type
proportions from test samples. The correlation plots can be displayed all
mixed or split based on cell type (\code{CellType}) or the number of cell
types present in the sample (\code{nMix}). See \code{facet.by} argument and
examples for more information. Moreover, a correlation value selected by user
is displayed as annotation on the plots. See \code{corr} argument for
details.
}
\examples{
## correlations by cell type
corrExpPredPlot(
  object = DDLSSmallCompleted,
  facet.by = "CellType",
  color.by = "CellType",
  corr = "both"
)
## correlations of all samples mixed
corrExpPredPlot(
  object = DDLSSmallCompleted,
  facet.by = NULL,
  color.by = "CellType",
  corr = "ccc",
  pos.x.label = 0.2,
  alpha.point = 0.3
)
}
\seealso{
\code{\link{calculateEvalMetrics}} \code{\link{distErrorPlot}}
\code{\link{blandAltmanLehPlot}} \code{\link{barErrorPlot}}
}
