% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DDLSChungComp}
\alias{DDLSChungComp}
\title{\code{\link{DigitalDLSorter}} object with Chung et al. 2017 data set
(completed version)}
\format{
An object of class \code{DigitalDLSorter} of length 1.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75688}
}
\usage{
DDLSChungComp
}
\description{
\code{\link{DigitalDLSorter}} object containing the original data set used to
generate \code{breast.chung.generic} and \code{breast.chung.specific} models
in order to show some examples in vignette and documentation. This version is
equivalent to "DDLSChung", but it includes additional steps (Deep Neural
Network, evaluation metrics...).
}
\details{
For more information about the data set and the models, see
\code{breast.chung.generic} or \code{breast.chung.specific}.
}
\references{
Chung, W., Eum, H. H., Lee, H. O., Lee, K. M., Lee, H. B., Kim,
K. T., et al. (2017). Single-cell RNA-seq enables comprehensive tumour and
immune cell profiling in primary breast cancer. Nat. Commun. 8 (1), 15081.
doi: \url{10.1038/ncomms15081}.

Torroja, C. y Sánchez-Cabo, F. (2019). digitalDLSorter: A Deep Learning
algorithm to quantify immune cell populations based on scRNA-Seq data.
Frontiers in Genetics 10, 978. doi: \url{10.3389/fgene.2019.00978}
}
\keyword{datasets}
