% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{deconv.data}
\alias{deconv.data}
\alias{deconv.data<-}
\title{Get and set \code{deconv.data} slot in a \code{DigitalDLSorter}
object.}
\usage{
deconv.data(object, name.data = NULL)

deconv.data(object, name.data = NULL) <- value
}
\arguments{
\item{object}{A \code{DigitalDLSorter} object.}

\item{name.data}{Name of the data. If it is \code{NULL} (by default),
all data contained in \code{deconv.data} slot are returned.}

\item{value}{A \code{list} whose names are the reference of the data stored.}
}
\description{
Get and set \code{deconv.data} slot in a \code{DigitalDLSorter}
object.
}
