% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{deconv.results}
\alias{deconv.results}
\alias{deconv.results<-}
\title{Get and set \code{deconv.results} slot in a
\code{\linkS4class{DigitalDLSorter}} object}
\usage{
deconv.results(object, name.data = NULL)

deconv.results(object, name.data = NULL) <- value
}
\arguments{
\item{object}{A \code{\linkS4class{DigitalDLSorter}} object.}

\item{name.data}{Name of the data. If it is \code{NULL} (by default), all
results contained in \code{deconv.results} slot are returned.}

\item{value}{A list whose names are the reference of the results
stored.}
}
\description{
Get and set \code{deconv.results} slot in a
\code{\linkS4class{DigitalDLSorter}} object
}
