% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnnModel.R
\name{deconvDigitalDLSorterObj}
\alias{deconvDigitalDLSorterObj}
\title{Deconvolute bulk gene expression samples (bulk RNA-Seq)}
\usage{
deconvDigitalDLSorterObj(
  object,
  name.data,
  batch.size = 128,
  normalize = TRUE,
  simplify.set = NULL,
  simplify.majority = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object with
\code{trained.data} and \code{deconv.data} slots.}

\item{name.data}{Name of the data stored in the \code{DigitalDLSorter}
object. If not provided, the first data set will be used.}

\item{batch.size}{Number of samples per gradient update. If not specified,
\code{batch.size} will default to 128.}

\item{normalize}{Normalize data before deconvolution (\code{TRUE} by
default).}

\item{simplify.set}{List specifying which cell types should be compressed
into a new label whose name will be the list item. See examples for
details. If provided, results are stored in a list with 'raw' and
'simpli.set' results.}

\item{simplify.majority}{List specifying which cell types should be
compressed into the cell type with the highest proportion in each sample.
Unlike \code{simplify.set}, it allows to maintain the complexity of the
results while compressing the information, as no new labels are created. If
provided, the results are stored in a list with 'raw' and 'simpli.majority'
results.}

\item{verbose}{Show informative messages during the execution.}
}
\value{
\code{\linkS4class{DigitalDLSorter}} object with
\code{deconv.results} slot. The resulting information is a data frame with
samples (\eqn{i}) as rows and cell types (\eqn{j}) as columns. Each entry
represents the proportion of \eqn{j} cell type in \eqn{i} sample. If
\code{simplify.set} or/and \code{simpplify.majority} are provided, the
\code{deconv.results} slot will contain a list with raw and simplified
results.
}
\description{
Deconvolute bulk gene expression samples (bulk RNA-Seq). This function
requires a \code{DigitalDLSorter} object with a trained Deep Neural Network
model (\code{\link{trained.model}} slot) and the new bulk RNA-Seq samples to
be deconvoluted in the \code{deconv.data} slot. See
\code{?\link{loadDeconvData}} for more details.
}
\details{
This function is intended for users who have built a devonvolution model
using their own single-cell RNA-Seq data. If you want to use a pre-trained
model to deconvolute your samples, see \code{?\link{deconvDigitalDLSorter}}.
}
\examples{
\dontrun{
if (!requireNamespace("digitalDLSorteRdata", quietly = TRUE)) {
  install.packages(
    "digitalDLSorteRdata", 
    repos = "https://diegommcc.github.io/digitalDLSorteRdataRepo/"
  )
}
library(digitalDLSorteRdata)
data(DDLSLi.list)
DDLSLi <- listToDDLS(DDLSLi.list)
data(TCGA.colon.se)
# to ensure compatibility
tensorflow::tf$compat$v1$disable_eager_execution()
# simplify arguments
simplify = list(Macrophages = c("Mc", "M"))
DDLSLi <- loadDeconvData(
  object = DDLSLi, data = TCGA.colon.se,
  name.data = "TCGA.colon"
)
DDLSLi <- deconvDigitalDLSorterObj(
  object = DDLSLi,
  name.data = "TCGA.colon",
  simplify.set = simplify,
  simplify.majority = simplify
)
}
}
\references{
Torroja, C. and Sánchez-Cabo, F. (2019). digitalDLSorter: A Deep
Learning algorithm to quantify immune cell populations based on scRNA-Seq
data. Frontiers in Genetics 10, 978. doi: \doi{10.3389/fgene.2019.00978}
}
\seealso{
\code{\link{trainDigitalDLSorterModel}}
\code{\linkS4class{DigitalDLSorter}}
}
