% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DigitalDLSorter-class}
\alias{DigitalDLSorter-class}
\alias{DigitalDLSorter}
\title{The DigitalDLSorter Class}
\description{
The DigitalDLSorter object is the core of \code{digitalDLSorteR}. This object
stores different intermediate data resulting from building new deconvolution
models from single-cell data. It is only used in the case of building new
deconvolution models. For deconvolving bulk samples using pre-trained models,
see \code{\link{deconvDigitalDLSorter}} function.
}
\details{
This object uses other classes to store the different types of data produced
during the process: \itemize{ \item \code{\linkS4class{SingleCellExperiment}}
class for single-cell RNA-seq data, using sparse matrix from the
\pkg{Matrix} package (\code{\linkS4class{dgCMatrix}} class) or
\code{HDF5Array} class in the case of using HDF5 files as
back-end (see below for more information). \item
\code{\linkS4class{ZINBParams}} class with estimated parameters for the
simulation of new single-cell profiles. \item
\code{\linkS4class{SummarizedExperiment}} class for storing bulk RNA-seq
data. \item \code{\linkS4class{ProbMatrixCellTypes}} class for the
composition cell matrices built during the process. See
\code{?\linkS4class{ProbMatrixCellTypes}} for details. \item
\code{\linkS4class{DigitalDLSorterDNN}} class to store the information
related to Deep Neural Network. This step is performed using \code{keras}.
See \code{?\linkS4class{DigitalDLSorterDNN}} for details. }
}
\section{Slots}{

\describe{
\item{\code{single.cell.real}}{Real single-cell data stored in a
\code{SingleCellExperiment} object. The count matrix is stored as
\code{\linkS4class{dgCMatrix}} or \code{HDF5Array} objects.}

\item{\code{zinb.params}}{\code{\linkS4class{ZINBParams}} object with estimated
parameters for the simulation of new single-cell expression profiles.}

\item{\code{single.cell.simul}}{Simulated single-cell expression profiles from
ZINB-WaVE model.}

\item{\code{prob.cell.types}}{\code{\linkS4class{ProbMatrixCellTypes}} class with
cell composition matrices built for the simulation of bulk RNA-seq profiles
with known cell composition.}

\item{\code{bulk.simul}}{A list with train and test simulated bulk RNA-seq samples.
Each entry is a \code{\linkS4class{SummarizedExperiment}} object. The
count matrices can be stored as \code{HDF5Array} files using
HDF5 files as back-end in case of RAM memory limitations.}

\item{\code{trained.model}}{\code{\linkS4class{DigitalDLSorterDNN}} object with all
the information related to the trained model. See
\code{?\linkS4class{DigitalDLSorterDNN}} for more details.}

\item{\code{deconv.data}}{List of \code{\linkS4class{SummarizedExperiment}} objects
where it is possible to store new bulk RNA-seq experiments for their
deconvolution. The name of the entries corresponds to the name of provided
data. See \code{\link{deconvDigitalDLSorterObj}} for details.}

\item{\code{deconv.results}}{Slot containing the deconvolution results from applying
the deconvolution model to data present in \code{\link{deconv.data}} slot.
It is a list in which the names corresponds to the data from which they
come.}

\item{\code{project}}{Name of the project.}

\item{\code{version}}{Version of DigitalDLSorteR this object was built under.

The package can be used in two ways: to build new deconvolution models from
single-cell RNA-seq data or to deconvolute bulk RNA-seq samples using
pre-trained models integrated into the package. If you want to build new
models, see \code{\link{loadSCProfiles}} function. If you want to use
pre-trained models, see \code{\link{deconvDigitalDLSorter}} function.

In order to provide a way of working with big amounts of data in machines
with RAM limitations, we provide the possibility of using HDF5 files as
back-end to store count matrices of both single-cell real/simulated and
bulk RNAseq profiles. To do this, we use \code{HDF5Array} and
\code{DelayedArray} classes from the homonymous packages.

Once the neural network is trained, it is possible to save it as RDS, RDA
and HDF5 files. Please, see \code{\linkS4class{DigitalDLSorterDNN}} for
more details.}
}}

