% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{eval.stats.model}
\alias{eval.stats.model}
\alias{eval.stats.model<-}
\title{Get and set \code{eval.stats.model} slot in a
\code{DigitalDLSorterDNN} object.}
\usage{
eval.stats.model(object)

eval.stats.model(object) <- value
}
\arguments{
\item{object}{A \code{DigitalDLSorterDNN} object.}

\item{value}{A \code{list} object with the resulting metrics after prediction
on test data with DNN model.}
}
\description{
Get and set \code{eval.stats.model} slot in a
\code{DigitalDLSorterDNN} object.
}
