% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{eval.stats.samples}
\alias{eval.stats.samples}
\alias{eval.stats.samples<-}
\title{Get and set \code{eval.stats.samples} slot in a
\code{DigitalDLSorterDNN} object.}
\usage{
eval.stats.samples(object, metrics = "All")

eval.stats.samples(object, metrics = "All") <- value
}
\arguments{
\item{object}{A \code{DigitalDLSorterDNN} object.}

\item{metrics}{Metrics to show (\code{'All'} by default)}

\item{value}{A \code{list} with evaluation metrics used for evaluating the
performance of the model over each sample from test data.}
}
\description{
Get and set \code{eval.stats.samples} slot in a
\code{DigitalDLSorterDNN} object.
}
