% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{exclusive.types}
\alias{exclusive.types}
\alias{exclusive.types,ProbMatrixCellTypes-method}
\alias{exclusive.types<-}
\alias{exclusive.types<-,ProbMatrixCellTypes-method}
\title{Get and set \code{exclusive.types} slot in a
\code{\linkS4class{ProbMatrixCellTypes}} object (deprecated)}
\usage{
exclusive.types(object)

exclusive.types(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{ProbMatrixCellTypes}} object.}

\item{value}{Optional slot that contains the exclusive cell types in the
experiment if they are provided. NULL by default.}
}
\description{
Get and set \code{exclusive.types} slot in a
\code{\linkS4class{ProbMatrixCellTypes}} object (deprecated)
}
