% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{features}
\alias{features}
\alias{features,DigitalDLSorterDNN-method}
\alias{features<-}
\alias{features<-,DigitalDLSorterDNN-method}
\title{Get and set \code{features} slot in a
\code{\linkS4class{DigitalDLSorterDNN}} object}
\usage{
features(object)

features(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorterDNN}} object.}

\item{value}{Vector with features (genes) considered by Deep Neural
Network model.}
}
\description{
Get and set \code{features} slot in a
\code{\linkS4class{DigitalDLSorterDNN}} object
}
