% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{final.data}
\alias{final.data}
\alias{final.data<-}
\title{Get and set \code{final.data} slot in a \code{DigitalDLSorter}
object.}
\usage{
final.data(object, type.data = "both")

final.data(object, type.data = "both") <- value
}
\arguments{
\item{object}{A \code{DigitalDLSorter} object.}

\item{type.data}{Element of the list. Can be 'train', 'test' or 'both' (the
last by default).}

\item{value}{A \code{list} with two elements, train and test, each one being
a \code{SummarizedExperiment} object with simulated bulk RNA-Seq samples
prepared for training. This samples have been normalized and shuffled.}
}
\description{
Get and set \code{final.data} slot in a \code{DigitalDLSorter}
object.
}
