% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBulk.R
\name{generateBulkCellMatrix}
\alias{generateBulkCellMatrix}
\title{Generate training and test cell composition matrices}
\usage{
generateBulkCellMatrix(
  object,
  cell.ID.column,
  cell.type.column,
  prob.design,
  num.bulk.samples,
  n.cells = 100,
  train.freq.cells = 2/3,
  train.freq.bulk = 2/3,
  proportions.train = c(10, 5, 20, 15, 10, 40),
  proportions.test = c(10, 5, 20, 15, 10, 40),
  balanced.type.cells = FALSE,
  exclusive.types = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object with
\code{single.cell.real} slot and, optionally, with \code{single.cell.simul}
slot.}

\item{cell.ID.column}{Name or number of the column in cells metadata
corresponding to cell names in expression matrix.}

\item{cell.type.column}{Name or number of the column in cells metadata
corresponding to the cell type of each cell.}

\item{prob.design}{\code{data.frame} with the frequency ranges expected for
each cell type present in the experiment. This information can be estimated
from literature or from the single-cell experiment itself. This
\code{data.frame} must be built by three columns with specific headers (see
Examples): \itemize{ \item A cell type column with the same name of the
cell type column in cells.metadata. If the name of the column is not the
same, function returns an error. Cell types must appear in cells.metadata.
\item A second column named 'from' with the start frequency for each cell
type. \item A third column named 'to' with the final frequency for each
cell type.}}

\item{num.bulk.samples}{Number of bulk samples proportions (and, therefore,
simulated bulk RNA-seq samples) that will be generated taking into account
training and test data. We recommend seting this value according to the
number of single-cell profiles available in
\code{\linkS4class{DigitalDLSorter}} object avoiding an excesive
re-sampling, but generating a big number of samples for a better training.}

\item{n.cells}{Number of cells that will be aggregated in order to simulate
one bulk RNA-seq sample (100 by default).}

\item{train.freq.cells}{Proportion of cells used for training set (2/3 by
default).}

\item{train.freq.bulk}{Proportion of bulk samples used for training set (2/3
by default).}

\item{proportions.train}{Vector of five integers that determine the
proportions of bulk samples generated by the different methods (see Details
or Torroja and Sánchez-Cabo, 2019 for more information). This vector
represents proportions, so they must add up 100. By default, a majority of
random samples without using predefined ranges will be generated.}

\item{proportions.test}{\code{proportions.train} for test samples.}

\item{balanced.type.cells}{Boolean indicating if training and test cells will
be split in a balanced way considering cell types (\code{FALSE} by
default).}

\item{exclusive.types}{Vector of cell types which allows to establish cell
types that biologically do not make sense to be mixed during the generation
of bulk samples. Some samples presents this exclusive cell types. If it is
equal to NULL (by default), all cell types will be mixed when generating
bulk samples (this argument will disappear).}

\item{verbose}{Show informative messages during the execution (\code{TRUE} by
default).}
}
\value{
A \code{\linkS4class{DigitalDLSorter}} object with
\code{prob.cell.types} slot containing a \code{list} with two
\code{\linkS4class{ProbMatrixCellTypes}} objects (train and test). For more
information about the structure of this class, see
\code{?\linkS4class{ProbMatrixCellTypes}}. The most important element is
the cell composition matrix, which is composed by \eqn{n} rows (being
\eqn{n} the number of bulk samples that will be generated) and \eqn{k}
columns (being \eqn{k} the number of cell types present in the experiment).
}
\description{
Generate training and test cell composition matrices for the simulation of
bulk RNA-seq samples with known cell composition using single-cell expression
profiles. The resulting \code{\linkS4class{ProbMatrixCellTypes}} object
contains the matrix which determines the proportion of the different cell
types that will compose the simulated bulk samples plus other information
relevant to the process. This function does not simulate bulk samples, this
task is performed by \code{\link{simBulkSamples}} or
\code{\link{trainDigitalDLSorterModel}} functions (see Documentation).
}
\details{
First, available single-cell profiles are split into training and test
subsets (2/3 for training and 1/3 for test by default (see
\code{train.freq.cells})) in order to avoid falsifying the results during the
evaluation of the model. Then, \code{num.bulk.samples} bulk samples
proportions are built and single-cell profiles that will be used for
simulating bulk RNA-seq samples are set, being 100 cells per bulk sample by
default (see \code{n.cells} argument). The proportions of training and test
bulk samples are set by \code{train.freq.bulk} (2/3 for training and 1/3 for
test by default). Finally, in order to avoid biases due to the composition of
bulk samples, proportions for the mixtures (bulk samples) of cell types
(\eqn{w_1,...,w_k}, where \eqn{k} is the number of cell types available in
single-cell profiles) are randomly generated by using five different
approaches:

\enumerate{ \item Cell proportions are randomly sampled from a truncated
uniform distribution with predefined limits according to a priori knowledge
of the abundance of each cell type (see \code{prob.design} argument). This
information can be inferred from the single cell analysis itself or from the
literature. \item A second set is generated by randomly permuting cell type
labels from a distribution generated by the previous method. \item Cell
proportions are randomly sampled as by method 1 without replacement. \item
Using the last method for generating proportions, cell types labels are
randomly sampled. \item Cell proportions are randomly sampled from a
Dirichlet distribution.}

If you want to inspect the distribution of cell type proportions generated by
each method during the process, they can be accessed by
\code{\link{showProbPlot}} function (see examples).
}
\examples{
## generate a data.frame with frequency ranges of each cell type
probMatrix <- data.frame(
  Cell_type = c("ER+", "HER2+", "ER+ and HER2+", "TNBC",
                "Stromal", "Monocyte", "Tme", "BGC",
                "Bmem", "DC", "Macrophage", "TCD8", "Treg"),
  from = c(rep(30, 4), 1, rep(0, 8)),
  to = c(rep(70, 4), 50, rep(15, 8))
)

DDLSSmallCompleted <- generateTrainAndTestBulkProbMatrix(
  object = DDLSSmallCompleted,
  cell.type.column = "Cell_type",
  prob.design = probMatrix,
  num.bulk.samples = 200,
  verbose = TRUE
)

}
\references{
Torroja, C. y Sánchez-Cabo, F. (2019). digitalDLSorter: A Deep
Learning algorithm to quantify immune cell populations based on scRNA-Seq
data. Frontiers in Genetics 10, 978. doi:
\href{https://doi.org/10.3389/fgene.2019.00978}{10.3389/fgene.2019.00978}
}
\seealso{
\code{\link{simBulkProfiles}}
\code{\linkS4class{ProbMatrixCellTypes}}
}
