% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{grid.search}
\alias{grid.search}
\alias{grid.search,DigitalDLSorter-method}
\alias{grid.search<-}
\alias{grid.search<-,DigitalDLSorter-method}
\title{Get and set \code{grid.search} slot in a
\code{\linkS4class{DigitalDLSorter}} object}
\usage{
grid.search(object)

grid.search(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object.}

\item{value}{List with two elements, train and test, each one being
a \code{\linkS4class{SummarizedExperiment}} object with simulated bulk
RNA-Seq samples.}

\item{type.data}{Element of the list. Can be \code{'train'}, \code{'test'} or
\code{'both'} (the last by default).}
}
\description{
Get and set \code{grid.search} slot in a
\code{\linkS4class{DigitalDLSorter}} object
}
