% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearch.R
\name{gridMetricDist}
\alias{gridMetricDist}
\title{Histogram of the distribution of a given metric}
\usage{
gridMetricDist(
  object,
  metric = "loss",
  set = "test.results",
  normalize = TRUE,
  theme = NULL,
  ...
)
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object with
\code{grid.search} slot.}

\item{metric}{The metric to plot. The available metrics can be known using
\code{metrics()}.}

\item{set}{The data set to use. Can be set to \code{test.results} (evaluation
using the test set) or \code{train.results} (data about the last training
epoch, with train and validation metrics).}

\item{normalize}{Boolean indicating if the metric is going to be normalized.}

\item{theme}{\pkg{ggplot2} theme.}
}
\description{
Display the distribution of a given metric between all generated models.
}
