% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearch.R
\name{gridMetricSDist}
\alias{gridMetricSDist}
\title{Histograms of the distribution of metrics}
\usage{
gridMetricSDist(
  object,
  set = "test.results",
  normalize = TRUE,
  theme = NULL,
  ...
)
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object with
\code{grid.search} slot.}

\item{set}{The data set to use. Can be set to \code{test.results} (evaluation
using the test set) or \code{train.results} (data about the last training
epoch, with train and validation metrics).}

\item{normalize}{Boolean indicating if the metrics are going to be
normalized.}

\item{theme}{\pkg{ggplot2} theme.}
}
\description{
Display the distribution of all metrics between all generated models.
}
