% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearch.R
\name{gridParamSDist}
\alias{gridParamSDist}
\title{Histograms of the distribution of parameters}
\usage{
gridParamSDist(object, theme = NULL, ...)
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object with
\code{grid.search} slot.}

\item{theme}{\pkg{ggplot2} theme.}

\item{...}{}
}
\description{
Display the distribution of all parameters between all generated models.
}
