% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{installPythonDepends}
\alias{installPythonDepends}
\title{Install Python dependencies for digitalDLSorteR}
\usage{
installPythonDepends(
  conda = "auto",
  install.conda = FALSE,
  miniconda.path = NULL
)
}
\arguments{
\item{conda}{Path to a conda executable. Use \code{"auto"} (by default)
allows \pkg{reticulate} to automatically find an appropriate conda binary.}

\item{install.conda}{Boolean indicating if install miniconda automatically
using \pkg{reticulate}. If \code{TRUE}, \code{conda} argument is ignored.
\code{FALSE} by default.}

\item{miniconda.path}{If \code{install.conda} is \code{TRUE}, you can set the
path where miniconda will be installed. If \code{NULL}, conda will find
automatically the proper place.}
}
\description{
This is a helper function to install Python dependencies needed: a Python
interpreter with TensorFlow Python library and its dependencies. It is
performed using the \pkg{reticulate} package and the installer of the
\pkg{tensorflow} R package. The available options are virtual or conda
environments. The are named digitaldlsorter. In any case, it can be manually
done as it is explained in
\url{https://diegommcc.github.io/digitalDLSorteR/articles/kerasIssues.html}.
}
\details{
This function is intended to make easier the installation of the requirements
needed to use \pkg{digitalDLSorteR}. It will automatically install Miniconda
(if wanted, see Parameters) and create an environment named
'digitaldlsorter-env'. If you want to use other python/conda environment, see
?tensorflow::use_condaenv and/or vignette('kerasIssues').
}
\examples{
\dontrun{
notesInstallation <- installPythonDepends(
  method = "auto", conda = "auto", install.conda = TRUE
)
}

}
