% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listToDDLS}
\alias{listToDDLS}
\title{Transform DigitalDLSorter-like list into an actual DigitalDLSorterDNN object}
\usage{
listToDDLS(listTo)
}
\arguments{
\item{listTo}{A list in which each element must correspond to each slot of an
\code{DigitalDLSorter} object. The names must be the same as the slot
names.}
}
\value{
\code{DigitalDLSorter} object the data provided in the original list.
}
\description{
Transform DigitalDLSorter-like list into an actual \code{DigitalDLSorter}
object. This function allows to generate the examples and the vignettes of
\pkg{digitalDLSorteR} package as these are the data used. These data are
stored in the digitalDLSorteRdata package.
}
\examples{
\dontrun{
if (!requireNamespace("digitalDLSorteRdata", quietly = TRUE)) {
  install.packages("digitalDLSorteRdata")
}
library(digitalDLSorteRdata)
data("DDLSLi.list")
DDLSLiObj <- listToDDLS(DDLSLi.list)
DDLSLiObj
}
  
}
\seealso{
\code{\link{listToDDLSDNN}}
}
