% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listToDDLSDNN}
\alias{listToDDLSDNN}
\title{Transform DigitalDLSorterDNN-like list into an actual DigitalDLSorterDNN
object}
\usage{
listToDDLSDNN(listTo)
}
\arguments{
\item{listTo}{A list in which each element must correspond to each slot of an
\code{DigitalDLSorterDNN} object. The names must be the same as the slot
names.}
}
\value{
\code{DigitalDLSorterDNN} object the data provided in the original
list.
}
\description{
Transform DigitalDLSorterDNN-like list into an actual
\code{DigitalDLSorterDNN} object. This function allows to use pre-trained
models in the \pkg{digitalDLSorteR} package.
}
\examples{
\dontrun{
if (requireNamespace("digitalDLSorteR", quietly = TRUE)) {
  data("colorectal.li")
  DDLSDNNObj <- listToDDLSDNN(colorectal.li)
  DDLSDNNObj
}
}
}
\seealso{
\code{\link{listToDDLS}}
}
