% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnnModel.R
\name{loadDeconvDataFromSummarizedExperiment}
\alias{loadDeconvDataFromSummarizedExperiment}
\title{Load data to deconvolute from \code{SummarizedExperiment} object.}
\usage{
loadDeconvDataFromSummarizedExperiment(object, se.object, name.data = NULL)
}
\arguments{
\item{object}{\code{DigitalDLSorter} object with \code{trained.model} slot.}

\item{se.object}{\code{SummarizedExperiment} object.}

\item{name.data}{Name with which the data is stored in \code{DigitalDLSorter}
object.}
}
\description{
Load data in \code{\link{DigitalDLSorter}} object to deconvolute from
\code{SummarizedExperiment} object.
}
\seealso{
\code{\link{trainDigitalDLSorterModel}}
\code{\link{deconvDigitalDLSorterObj}}
}
