% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearch.R
\name{loadGridSearchRDS}
\alias{loadGridSearchRDS}
\title{Load the grid.search slot from a RDS file}
\usage{
loadGridSearchRDS(object, file)
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}}}

\item{file}{A string indicating the path and file name to load the data from.}
}
\description{
Load a RDS file containing the \code{grid.search} slot of a
\code{\linkS4class{DigitalDLSorter}} object.
}
