% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadTrainedModelFromH5}
\alias{loadTrainedModelFromH5}
\title{Load from HDF5 file a trained DigitalDLSorter DNN model.}
\usage{
loadTrainedModelFromH5(object, file.path, reset.slot = FALSE)
}
\arguments{
\item{object}{\code{\link{DigitalDLSorter}} object with \code{trained.model}
slot.}

\item{file.path}{Valid file path where model are stored.}

\item{reset.slot}{Remove \code{trained.slot} if it already exists. A new
\code{\link{DigitalDLSorterDNN}} object will be formed, but it will not
contain other slots (\code{FALSE} by default).}
}
\description{
Load from HDF5 file a trained DigitalDLSorter DNN model into a
\code{\link{DigitalDLSorter}} object. Note that HDF5 file must be a keras
valid trained model.
}
\seealso{
\code{\link{trainDigitalDLSorterModel}}
\code{\link{deconvDigitalDLSorterObj}}
\code{\link{saveTrainedModelAsH5}}
}
