% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{method}
\alias{method}
\alias{method,ProbMatrixCellTypes-method}
\alias{method<-}
\alias{method<-,ProbMatrixCellTypes-method}
\title{Get and set \code{method} slot in a
\code{\linkS4class{ProbMatrixCellTypes}} object}
\usage{
method(object)

method(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{ProbMatrixCellTypes}} object.}

\item{value}{Vector with names of cells present in the object.}
}
\description{
Get and set \code{method} slot in a
\code{\linkS4class{ProbMatrixCellTypes}} object
}
