% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{model}
\alias{model}
\alias{model,DigitalDLSorterDNN-method}
\alias{model<-}
\alias{model<-,DigitalDLSorterDNN-method}
\title{Get and set \code{model} slot in a
\code{\linkS4class{DigitalDLSorterDNN}} object}
\usage{
model(object)

model(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorterDNN}} object.}

\item{value}{\code{keras.engine.sequential.Sequential} object with a
trained Deep Neural Network model.}
}
\description{
Get and set \code{model} slot in a
\code{\linkS4class{DigitalDLSorterDNN}} object
}
