% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{predict.results}
\alias{predict.results}
\alias{predict.results<-}
\title{Get and set \code{predict.results} slot in a \code{DigitalDLSorterDNN}
object.}
\usage{
predict.results(object)

predict.results(object) <- value
}
\arguments{
\item{object}{A \code{DigitalDLSorterDNN} object.}

\item{value}{A \code{matrix} object with prediction results on test data.}
}
\description{
Get and set \code{predict.results} slot in a \code{DigitalDLSorterDNN}
object.
}
