% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ProbMatrixCellTypes-class}
\alias{ProbMatrixCellTypes-class}
\alias{ProbMatrixCellTypes}
\title{The Class ProbMatrixCellTypes.}
\description{
The ProbMatrixCellTypes class is a data storage class that stores information
concerning to cell composition matrix used for the simulation of bulk
samples. This matrix corresponds with \code{prob.matrix} slot. The rest of
slots are additional information generated during the process and required
for subsequent steps.
}
\details{
As described in Torroja and Sanchez-Cabo, 2019, the proportions are built by
five different methods in order to avoid biases due to the composition of the
bulk samples. In \code{plots} slot, different representations of these
probabilities are stored with the aim of offering a method to monitor the
different sets of samples generated during the process. These plots can be
displayed with \code{\link{showProbPlot}} function. See documentation for
details.
}
\section{Slots}{

\describe{
\item{\code{prob.matrix}}{Matrix of cell proportions generated for the simulation of
bulk samples. Rows correspond with bulk samples which will be generated
(\eqn{i}), columns are the cell types present in single-cell data provided
(\eqn{j}) and each entry is the proportion of \eqn{j} cell type in \eqn{i}
sample.}

\item{\code{cell.names}}{Matrix in which names of cells that will compound each bulk
samples are stored.}

\item{\code{set.list}}{List of cells ordered according to the cell type to which they
belong.}

\item{\code{set}}{Vector with the names of cells present in the object.}

\item{\code{exclusive.types}}{Optional slot that contains the exclusive cell types on
the experiment if they are provided. NULL by default.}

\item{\code{plots}}{List of lists with plots showing the distribution of cell
proportions generated by each method during the process.}

\item{\code{type.data}}{Character with the type of data contained: training or test.}
}}

\references{
Torroja, C. y Sánchez-Cabo, F. (2019). digitalDLSorter: A Deep
Learning algorithm to quantify immune cell populations based on scRNA-Seq
data. Frontiers in Genetics 10, 978. doi: \url{10.3389/fgene.2019.00978}
}
