% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearch.R
\name{saveGridSearchRDS}
\alias{saveGridSearchRDS}
\title{Save the grid.search slot in a RDS file}
\usage{
saveGridSearchRDS(object, file)
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object with
\code{grid.search} slot.}

\item{file}{A string indicating the path and file name to store the data.}
}
\description{
Write a RDS file containing the \code{grid.search} slot of a
\code{\linkS4class{DigitalDLSorter}} object.
}
