% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{set.list}
\alias{set.list}
\alias{set.list<-}
\title{Get and set \code{set.list} slot in a \code{ProbMatrixCellTypes}
object.}
\usage{
set.list(object)

set.list(object) <- value
}
\arguments{
\item{object}{A \code{ProbMatrixCellTypes} object.}

\item{value}{List of cells ordered according to the cell type to which they
belong.}
}
\description{
Get and set \code{set.list} slot in a \code{ProbMatrixCellTypes}
object.
}
