% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{set}
\alias{set}
\alias{set<-}
\title{Get and set \code{set} slot in a
\code{\linkS4class{ProbMatrixCellTypes}} object.}
\usage{
set(object)

set(object) <- value
}
\arguments{
\item{object}{A \code{\linkS4class{ProbMatrixCellTypes}} object.}

\item{value}{Vector with the names of cells present in the object.}
}
\description{
Get and set \code{set} slot in a
\code{\linkS4class{ProbMatrixCellTypes}} object.
}
