% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showProbPlot}
\alias{showProbPlot}
\title{Show distribution plots of cell proportions generated by
\code{\link{generateTrainAndTestBulkProbMatrix}}.}
\usage{
showProbPlot(object, type.data, set, type.plot = "boxplot")
}
\arguments{
\item{object}{\code{DigitalDLSorter} object with \code{prob.cell.types} slot
with \code{plot} slot.}

\item{type.data}{Subset of data to show: \code{train} or \code{test}.}

\item{set}{Integer determining which of the 6 different approaches used for
generating the probability matrix show. Note that 1 and 2 follow the same
distribution.}

\item{type.plot}{Character determining which type of visualization show. It
can be \code{boxplot}, \code{violinplot}, \code{linesplot} or \code{nmix}.
See Description for more information.}
}
\value{
\code{ggplot} object.
}
\description{
Show distribution plots of cell proportions generated by the five different
methods used by \code{\link{generateTrainAndTestBulkProbMatrix}}. These
frequencies will determine the proportion of different cell types used during
the simulation of bulk RNA-Seq samples. There are 6 subsets of proportions (1
and 2 are generated by the same way) by different approaches that can be
shown with three types of visualizations: boxplot, violinplot and linesplot.
You also can represent the probabilities based on the number of different
cell types present in samples setting \code{type.plot = 'nMix'}.
}
\details{
These plots are only for diagnostic purposes. This is the reason because they
are generated without any parameter introduced by the user.
}
\examples{
lapply(1:6, function(x) {
  showProbPlot(DDLSSmallCompleted,
               type.data = "train",
               set = x,
               type.plot = "boxplot")
})

}
\seealso{
\code{\link{generateTrainAndTestBulkProbMatrix}}
}
