% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{single.cell.final}
\alias{single.cell.final}
\alias{single.cell.final<-}
\title{Get and set \code{single.cell.final} slot in a \code{DigitalDLSorter}
object.}
\usage{
single.cell.final(object)

single.cell.final(object) <- value
}
\arguments{
\item{object}{A \code{DigitalDLSorter} object.}

\item{value}{A \code{SingleCellExperiment} object with real and simulated
single-cell profiles.}
}
\description{
Get and set \code{single.cell.final} slot in a \code{DigitalDLSorter}
object.
}
