% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{single.cell.real}
\alias{single.cell.real}
\alias{single.cell.real,DigitalDLSorter-method}
\alias{single.cell.real<-}
\alias{single.cell.real<-,DigitalDLSorter-method}
\title{Get and set \code{single.cell.real} slot in a
\code{\linkS4class{DigitalDLSorter}} object}
\usage{
single.cell.real(object)

single.cell.real(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object.}

\item{value}{\code{\linkS4class{SingleCellExperiment}} object with real
single-cell profiles.}
}
\description{
Get and set \code{single.cell.real} slot in a
\code{\linkS4class{DigitalDLSorter}} object
}
