% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA.breast.small}
\alias{TCGA.breast.small}
\title{Breast cancer bulk RNA-Seq samples from TCGA Research Network}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 44831 rows and 15 columns.
}
\source{
\url{https://www.cancer.gov/tcga}
}
\usage{
TCGA.breast.small
}
\description{
Subset of breast cancer bulk RNA-Seq samples from TCGA Research Network.
FPKMs were transformed to TPMs and aggregated based on SYMBOL genes.
}
\references{
Koboldt, D. C., Fulton, R. S., McLellan, M. D., Schmidt, H.,
Kalicki-Veizer, J., McMichael, J. F., et al. (2012). Comprehensive
molecular portraits of human breast tumours. Nature 490 (7418), 61–70. doi:
10.1038/nature11412

Ciriello, G., Gatza, M. L., Beck, A. H., Wilkerson, M. D., Rhie, S. K.,
Pastore, A., et al. (2015). Comprehensive molecular portraits of invasive
lobular breast cancer. Cell. 163 (2), 506–519. doi:
10.1016/j.cell.2015.09.033

Torroja, C. y Sánchez-Cabo, F. (2019). digitalDLSorter: A Deep Learning
algorithm to quantify immune cell populations based on scRNA-Seq data.
Frontiers in Genetics 10, 978. doi: \url{10.3389/fgene.2019.00978}
}
\keyword{datasets}
