% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{test.pred}
\alias{test.pred}
\alias{test.pred<-}
\title{Get and set \code{test.pred} slot in a
\code{\linkS4class{DigitalDLSorterDNN}} object.}
\usage{
test.pred(object)

test.pred(object) <- value
}
\arguments{
\item{object}{A \code{\linkS4class{DigitalDLSorterDNN}} object.}

\item{value}{A \code{matrix} object with prediction results on test data.}
}
\description{
Get and set \code{test.pred} slot in a
\code{\linkS4class{DigitalDLSorterDNN}} object.
}
