% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{zinb.params}
\alias{zinb.params}
\alias{zinb.params,DigitalDLSorter-method}
\alias{zinb.params<-}
\alias{zinb.params<-,DigitalDLSorter-method}
\title{Get and set \code{zinb.params} slot in a
\code{\linkS4class{DigitalDLSorter}} object}
\usage{
zinb.params(object)

zinb.params(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object.}

\item{value}{\code{\linkS4class{ZinbParametersModel}} object with a valid
\code{ZinbModel} object.}
}
\description{
Get and set \code{zinb.params} slot in a
\code{\linkS4class{DigitalDLSorter}} object
}
