% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ZinbParametersModel-class}
\alias{ZinbParametersModel-class}
\alias{ZinbParametersModel}
\title{The Class ZinbParametersModel}
\description{
The ZinbParametersModel class is a wrapper class of the
\code{ZinbModel} class from zinbwave package.
}
\details{
This is a wrapper class of the \code{ZinbModel} class. It
consists of only one slot (\code{zinbwave.mode}) that contains the
\code{ZinbModel} object.
}
\section{Slots}{

\describe{
\item{\code{zinbwave.model}}{A valid \code{ZinbModel} object.}
}}

\references{
Risso, D., Perraudeau, F., Gribkova, S. et al. (2018). A general
and flexible method for signal extraction from single-cell RNA-seq data.
Nat Commun 9, 284. doi: \doi{10.1038/s41467-017-02554-5}.
}
