% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalMetrics.R
\name{barErrorPlot}
\alias{barErrorPlot}
\title{Generate bar error plot and its dispersion by cell types or by number of
different cell types in test bulk samples}
\usage{
barErrorPlot(
  object,
  error = "MSE",
  by = "CellType",
  dispersion = "se",
  filter.sc = TRUE,
  title = NULL,
  angle = NULL,
  theme = NULL
)
}
\arguments{
\item{object}{\code{DigitalDLSorter} object with \code{trained.model} slot
containing metrics in \code{test.deconv.metrics} slot.}

\item{error}{'MAE' or 'MSE.}

\item{by}{Variable used to display errors. Available options are:
'nCellTypes', 'CellType'.}

\item{dispersion}{Standard error (\code{'se'}) or standard deviation
(\code{'sd'}). The first by default.}

\item{filter.sc}{Boolean indicating if filter single-cell profiles and only
display correlations of results associated with bulk samples (\code{TRUE}
by default).}

\item{title}{Title of the plot.}

\item{angle}{Angle of ticks.}

\item{theme}{ggplot theme.}
}
\description{
Generate bar error plot and its dispersion by cell types (\code{CellType}) or
by number of different cell types (\code{nCellTypes}) in test bulk samples.
}
\examples{
barErrorPlot(
  object = DDLSChungComp,
  error = "MSE",
  by = "CellType"
)

barErrorPlot(
  object = DDLSChungComp,
  error = "MAE",
  by = "nCellTypes"
)
}
\seealso{
\code{\link{calculateEvalMetrics}} \code{\link{corrExpPredPlot}}
\code{\link{distErrorPlot}} \code{\link{blandAltmanLehPlot}}
}
