% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{barPlotCellTypes}
\alias{barPlotCellTypes}
\alias{barPlotCellTypes,DigitalDLSorter-method}
\alias{barPlotCellTypes,ANY-method}
\title{Bar plot with deconvoluted cell type proportions}
\usage{
barPlotCellTypes(
  data,
  colors = NULL,
  simplify = NULL,
  color.line = NA,
  x.label = "Bulk samples",
  rm.x.text = FALSE,
  title = "Results of deconvolution",
  legend.title = "Cell types",
  angle = 90,
  theme = NULL,
  ...
)

\S4method{barPlotCellTypes}{DigitalDLSorter}(
  data,
  colors = NULL,
  simplify = NULL,
  color.line = NA,
  x.label = "Bulk samples",
  rm.x.text = FALSE,
  title = "Results of deconvolution",
  legend.title = "Cell types",
  angle = 90,
  theme = NULL,
  name.data = NULL
)

\S4method{barPlotCellTypes}{ANY}(
  data,
  colors,
  color.line = NA,
  x.label = "Bulk samples",
  rm.x.text = FALSE,
  title = "Results of deconvolution",
  legend.title = "Cell types",
  angle = 90,
  theme = NULL
)
}
\arguments{
\item{data}{\code{\linkS4class{DigitalDLSorter}} object with
\code{deconv.results} slot or data frame/matrix with cell types as columns
and samples as rows.}

\item{colors}{Vector of colors that will be used.}

\item{simplify}{Type of simplification performed during deconvolution. It can
be \code{simpli.set} or \code{simpli.maj} (\code{NULL} by default). It is
only for \code{\linkS4class{DigitalDLSorter}} objects.}

\item{color.line}{Color of border bars.}

\item{x.label}{Label of x axis.}

\item{rm.x.text}{Logical value indicating if remove x axis ticks (name of
samples).}

\item{title}{Title of plot.}

\item{legend.title}{Title of legend plot.}

\item{angle}{Angle of text ticks.}

\item{theme}{\pkg{ggplot2} theme.}

\item{...}{Other arguments for specific methods.}

\item{name.data}{If a \code{\linkS4class{DigitalDLSorter}} is given, name of
the element that stores the results in \code{deconv.results} slot.}
}
\description{
This function allows to plot a bar plot with the deconvoluted cell type
proportions of a given bulk RNA-seq sample using ggplot2.
}
\examples{
# Using a matrix
\dontrun{barPlotCellTypes(deconvResults)}

# Using a DigitalDLSorter object
\dontrun{
  barPlotCellTypes(DDLSChung, name.data = "TCGA.breast")
}
}
\seealso{
\code{\link{deconvDigitalDLSorter}}
\code{\link{deconvDigitalDLSorterObj}}
}
