% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalMetrics.R
\name{blandAltmanLehPlot}
\alias{blandAltmanLehPlot}
\title{Generate Bland-Altman agreement plots between predicted and expected cell
type proportions from test data}
\usage{
blandAltmanLehPlot(
  object,
  colors,
  color.by,
  facet.by = NULL,
  log.2 = FALSE,
  filter.sc = TRUE,
  density = TRUE,
  color.density = "darkblue",
  size.point = 0.05,
  alpha.point = 1,
  ncol = NULL,
  nrow = NULL,
  title = NULL,
  theme = NULL,
  ...
)
}
\arguments{
\item{object}{\code{DigitalDLSorter} object with \code{trained.model} slot
containing metrics in \code{test.deconv.metrics} slot.}

\item{colors}{Vector of colors to use. Only vectors with a number of colors
equal to or greater than the levels of \code{color.by} will be accepted. By
default it is used a list of custom colors provided by the package.}

\item{color.by}{Variable used to color data. The options are
\code{nCellTypes} and \code{CellType}.}

\item{facet.by}{Variable used to display data in different panels. If it is
\code{NULL}, the plot is not separated into different panels. The options
are \code{nCellTypes} (by number of different cell types) and
\code{CellType} (by cell type).}

\item{log.2}{If show  Bland-Altman agreement plot in log2 space (\code{FALSE}
by default).}

\item{filter.sc}{Boolean indicating if filter single-cell profiles and only
display correlations of results associated with bulk samples (\code{TRUE}
by default).}

\item{density}{Boolean indicating if show density lines (\code{TRUE} by
default).}

\item{color.density}{Color of density lines if \code{density} argument is
equal to \code{TRUE}.}

\item{size.point}{Size of points (0.1 by default).}

\item{alpha.point}{Alpha of points (0.1 by default).}

\item{ncol}{Number of columns if \code{facet.by} is different than
\code{NULL}.}

\item{nrow}{Number of rows if \code{facet.by} is different than \code{NULL}.}

\item{title}{Title of the plot.}

\item{theme}{ggplot theme.}

\item{...}{Additional argument for \code{facet_wrap} ggplot function if
\code{facet.by} is not equal to \code{NULL}.}
}
\description{
Generate Bland-Altman agreement plots between predicted and expected cell
type proportions from test data. The Bland-Altman agreement plots can be
displayed all mixed or splited by cell type (\code{CellType}) or the number
of cell types present in the sample (\code{nCellTypes}). See \code{facet.by}
argument and examples for more information.
}
\examples{
## Bland-Altman plot by cell type
blandAltmanLehPlot(
  object = DDLSChungComp,
  facet.by = "CellType",
  color.by = "CellType"
)
## Bland-Altman plot of all samples mixed
blandAltmanLehPlot(
  object = DDLSChungComp,
  facet.by = NULL,
  color.by = "CellType",
  alpha.point = 0.3,
  log2 = TRUE
)
}
\seealso{
\code{\link{calculateEvalMetrics}} \code{\link{corrExpPredPlot}}
\code{\link{distErrorPlot}} \code{\link{barErrorPlot}}
}
