% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{bulk.simul}
\alias{bulk.simul}
\alias{bulk.simul,DigitalDLSorter-method}
\alias{bulk.simul<-}
\alias{bulk.simul<-,DigitalDLSorter-method}
\title{Get and set \code{bulk.simul} slot in a
\code{\linkS4class{DigitalDLSorter}} object}
\usage{
bulk.simul(object, type.data = "both")

bulk.simul(object, type.data = "both") <- value
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object.}

\item{type.data}{Element of the list. Can be \code{'train'}, \code{'test'} or
\code{'both'} (the last by default).}

\item{value}{List with two elements, train and test, each one being
a \code{SummarizedExperiment} object with simulated bulk
RNA-Seq samples.}
}
\description{
Get and set \code{bulk.simul} slot in a
\code{\linkS4class{DigitalDLSorter}} object
}
