% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalMetrics.R
\name{calculateEvalMetrics}
\alias{calculateEvalMetrics}
\title{Calculate evaluation metrics for bulk RNA-seq samples from test data.}
\usage{
calculateEvalMetrics(object, metrics = c("MAE", "MSE"))
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object with
\code{single.cell.final} and \code{\linkS4class{DigitalDLSorterDNN}} slots.}

\item{metrics}{Metrics used for evaluating the performance of the model. Mean
absolute error (MAE) and mean squared error (MSE) by default.}
}
\value{
A \code{\linkS4class{DigitalDLSorterDNN}} object with
\code{trained.model} slot containing a
\code{\linkS4class{DigitalDLSorterDNN}} object with
\code{test.deconv.metrics} slot.
}
\description{
Calculate evaluation metrics for bulk RNA-seq samples from test data in order
to know the performance of the model. By default, absolute error (AbsErr),
proportional absolute error (ppAbsErr), squared error (SqrErr) and
proportional squared error (ppSqrErr) are calculated for each test sample.
Moreover, each one of these metrics is aggregated using their mean values by
three criteria: each cell type (\code{CellType}), probability bins in ranges
of 0.1 (\code{pBin}) and number of different cell types present in the sample
\code{}. Finally, the process is repeated only for bulk samples, removing
single-cell profiles from the evaluation. Evaluation metrics are available in
\code{test.deconv.metrics} slot of \code{\linkS4class{DigitalDLSorterDNN}}
object (\code{trained.model} slot of \code{\linkS4class{DigitalDLSorter}}
object).
}
\examples{
if (requireNamespace("digitalDLSorteRdata", quietly = TRUE)) {
  library(digitalDLSorteRdata)
  data(DDLSLiComp)
  DDLSLiComp <- calculateEvalMetrics(
    object = DDLSLiComp
  )
}
}
\seealso{
\code{\link{distErrorPlot}} \code{\link{corrExpPredPlot}}
\code{\link{blandAltmanLehPlot}} \code{\link{barErrorPlot}}
}
