% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{deconv.data}
\alias{deconv.data}
\alias{deconv.data,DigitalDLSorter-method}
\alias{deconv.data<-}
\alias{deconv.data<-,DigitalDLSorter-method}
\title{Get and set \code{deconv.data} slot in a
\code{\linkS4class{DigitalDLSorter}} object}
\usage{
deconv.data(object, name.data = NULL)

deconv.data(object, name.data = NULL) <- value
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object.}

\item{name.data}{Name of data. If it is \code{NULL} (by default), all
data contained in \code{deconv.data} slot are returned.}

\item{value}{List whose names are the reference of the data stored.}
}
\description{
Get and set \code{deconv.data} slot in a
\code{\linkS4class{DigitalDLSorter}} object
}
