% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnnModel.R
\name{deconvDigitalDLSorterObj}
\alias{deconvDigitalDLSorterObj}
\title{Deconvolute bulk gene expression samples (bulk RNA-Seq)}
\usage{
deconvDigitalDLSorterObj(
  object,
  name.data,
  batch.size = 128,
  normalize = TRUE,
  simplify.set = NULL,
  simplify.majority = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{\code{\link{DigitalDLSorter}} object with \code{trained.data}
and \code{deconv.data} slots.}

\item{name.data}{Name of the data store in \code{DigitalDLSorter} object. If
it is not provided, the first data set will be used.}

\item{batch.size}{Number of samples per gradient update. If unspecified,
\code{batch.size} will default to 128.}

\item{normalize}{Normalize data before deconvolution. \code{TRUE} by default.}

\item{simplify.set}{List specifying which cell types should be compressed
into a new label whose name will be the list item. See examples for
details. The results are stored in a list with normal and simpli.majority
results (if provided). The name of the element in the list is
\code{'simpli.set'}.}

\item{simplify.majority}{List specifying which cell types should be
compressed into the cell types with greater proportion in each sample.
Unlike \code{simplify.set}, it allows to maintain the complexity of the
results while compressing the information, because it is not created a new
label. The results are stored in a list with normal and simpli.set results
(if provided). The name of the element in the list is
\code{'simpli.majority'}.}

\item{verbose}{Show informative messages during the execution.}
}
\value{
A \code{data.frame} with samples (\eqn{i}) as rows and cell types
(\eqn{j}) as columns. Each entry represents the proportion of \eqn{j} cell
type in \eqn{i} sample.
}
\description{
Deconvolute bulk gene expression samples (bulk RNA-Seq) enumerating and
quantifying the proportion of cell types present in a bulk sample. This
function needs a \code{DigitalDLSorter} object with a trained DNN model
(\code{\link{trained.model}} slot) and bulk samples for deconvoluting in
\code{deconv.data} slot.
}
\details{
This function is oriented for users that have trained a DNN model using their
own data. If you want to use a pre-trained model, see
\code{\link{deconvDigitalDLSorter}}.
}
\examples{
## to ensure compatibility
\dontrun{
tensorflow::tf$compat$v1$disable_eager_execution()
## simplify arguments
simplify <- list(Tumor = c("ER+", "HER2+", "ER+ and HER2+", "TNBC"),
                 Bcells = c("Bmem", "BGC"))

## all results are stored in DigitalDLSorter object
DDLSSmallCompleted <- deconvDigitalDLSorterObj(
  object = DDLSSmallCompleted,
  name.data = "TCGA.breast",
  simplify.set = simplify,
  simplify.majority = simplify
)
}
}
\references{
Torroja, C. and Sánchez-Cabo, F. (2019). digitalDLSorter: A Deep
Learning algorithm to quantify immune cell populations based on scRNA-Seq
data. Frontiers in Genetics 10, 978. doi:
\href{https://doi.org/10.3389/fgene.2019.00978}{10.3389/fgene.2019.00978}
}
\seealso{
\code{\link{trainDigitalDLSorterModel}}
\code{\link{DigitalDLSorter}}
}
