% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DigitalDLSorter-class}
\alias{DigitalDLSorter-class}
\alias{DigitalDLSorter}
\title{The DigitalDLSorter Class.}
\description{
The DigitalDLSorter object is the core of digitalDLSorteR. This object stores
the different intermediate data resulting from running pipeline from real
single-cell data to the trained Deep Neural Network, including the data on
which to carry out the process of devonvolution. Only it is used in the case
of building new deconvolution models. For deconvoluting bulk samples using
pre-trained models, see \code{\link{deconvDigitalDLSorter}} function.
}
\details{
This object uses other classes to store the different type of data produced
during the process: \itemize{ \item \code{SingleCellExperiment} class for
single-cell RNA-seq data, using sparse matrix from the \code{Matrix} package
(\code{dgCMatrix} class) to store the matrix of counts. \item
\code{ZinbParams} class with the estimated parameters for the simulation of
new single-cell profiles. \item \code{SummarizedExperiment} class for storing
bulk RNA-seq data. In this case, it is possible to load all data in memory or
the use of HDF5 files as back-end by \code{DelayedArray} and \code{HDF5Array}
packages. See \code{\link{generateBulkSamples}} for details. \item
\code{\link{ProbMatrixCellTypes}} class for the composition cell matrices
built during the process. See \code{?ProbMatrixCellTypes} for details. \item
\code{\link{DigitalDLSorterDNN}} class for storing the trained Deep Neural
Network. This step is performed by \code{keras}. See
\code{\link{DigitalDLSorterDNN}} for details. }
}
\section{Slots}{

\describe{
\item{\code{single.cell.real}}{Real single-cell data stored in a
\code{SingleCellExperiment} object. The counts matrix is stored as a
\code{dgCMatrix} object to optimize the amount of used memory.}

\item{\code{zinb.params}}{\code{ZinbParams} object with estimated parameters for the
simulation of new single-cell expression profiles.}

\item{\code{single.cell.simul}}{Simulated single-cell expression profiles used for
simulating bulk RNA-seq profiles with known cell composition.}

\item{\code{prob.cell.types}}{\code{ProbMatrixCellTypes} class with the cell
composition  matrix built for the simulation of bulk RNA-seq profiles. The
entries determine the proportion of single-cell types that will constitute
the simulated bulk samples.}

\item{\code{bulk.simul}}{A list with two elements: train and test simulated bulk
RNA-seq. This data are stored as a \code{SummarizedExperiment} object. We
recommend the use of HDF5 file as a back-end due to the large amount of
memory that they occupy.}

\item{\code{trained.model}}{\code{\link{DigitalDLSorterDNN}} object with the trained
model, different metrics obtained during the training and evaluation
metrics from the application of the model on test data. After executing
\code{\link{calculateEvalMetrics}}, it is alto possible to find the results
of the model evaluation.}

\item{\code{deconv.data}}{Optional slot where is possible to store new bulk samples
for its deconvolution. It is a list whose name is the name of the data
provided. It is possible to store more than one dataset to make
predictions. See \code{\link{deconvDigitalDLSorterObj}} for details.}

\item{\code{deconv.results}}{Slot where the results from the deconvolution process
over \code{\link{deconv.data}} data are stored. It is a list whose name is
the name of the data from which they come.}

\item{\code{project}}{Name of the project.}

\item{\code{version}}{Version of DigitalDLSorteR this object was built under.

The package can be used in two ways: to build new models of deconvolution
from scRNA-seq data or to deconvolute bulk RNA-seq samples using
pre-trtained models integrated into the package. If you want to build new
models, see \code{\link{loadRealSCProfiles}} or
\code{\link{loadFinalSCProfiles}} functions. If yoy want to use pre-trained
models, see \code{\link{deconvDigitalDLSorter}} function.}
}}

