% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalMetrics.R
\name{distErrorPlot}
\alias{distErrorPlot}
\title{Generate box plot or violin plot to show how errors are distributed}
\usage{
distErrorPlot(
  object,
  error,
  colors,
  x.by = "pBin",
  facet.by = NULL,
  color.by = "nCellTypes",
  filter.sc = TRUE,
  error.label = FALSE,
  pos.x.label = 4.6,
  pos.y.label = NULL,
  size.point = 0.1,
  alpha.point = 1,
  type = "violinplot",
  ylimit = NULL,
  nrow = NULL,
  ncol = NULL,
  title = NULL,
  theme = NULL,
  ...
)
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object with
\code{trained.model} slot containing metrics in \code{test.deconv.metrics}
slot of \code{\linkS4class{DigitalDLSorterDNN}} object.}

\item{error}{Which error is going to be represented. Available errors are
absolute error (\code{'AbsErr'}), proportional absolute error
(\code{'ppAbsErr'}), squared error (\code{'SqrErr'}) or proportional
squared error (\code{'ppSqrErr'}).}

\item{colors}{Vector of colors to use. Only vectors with a number of colors
equal to or greater than the levels of \code{color.by} will be accepted. By
default a list of custom colors provided by the package is used.}

\item{x.by}{Variable used for X axis. When \code{facet.by} is not
\code{NULL}, the best option is \code{pBin} (probability bins). The options
are \code{nCellTypes} (by number of different cell types), \code{CellType}
(by cell type) and \code{pBin}.}

\item{facet.by}{Variable used to display data in different panels. If it is
\code{NULL}, the plot is not separated into different panels. The options
are \code{nCellTypes} (by number of different cell types) and
\code{CellType} (by cell type).}

\item{color.by}{Variable used to color data. The options are
\code{nCellTypes} and \code{CellType}.}

\item{filter.sc}{Boolean indicating whether to filter single-cell profiles
and only display errors associated with bulk samples (\code{TRUE} by
default).}

\item{error.label}{Boolean indicating if to show average error as annotation
in plots (\code{FALSE} by default).}

\item{pos.x.label}{Position on the X axis of the errors annotations.}

\item{pos.y.label}{Position on the Y axis of the errors annotations.}

\item{size.point}{Size of points (0.1 by default).}

\item{alpha.point}{Alpha of points (0.1 by default).}

\item{type}{Type of plot: \code{'boxplot'} or \code{'violinplot'}. The latter
by default.}

\item{ylimit}{Upper limit in y axis if it is required (\code{NULL} by
default).}

\item{nrow}{Number of rows if \code{facet.by} is different from \code{NULL}.}

\item{ncol}{Number of columns if \code{facet.by} is different from
\code{NULL}.}

\item{title}{Title of the plot.}

\item{theme}{\pkg{ggplot2} theme.}

\item{...}{Additional arguments for \link[ggplot2]{facet_wrap} \pkg{ggplot2}
function if \code{facet.by} is not equal to \code{NULL}.}
}
\value{
ggplot2 plot
}
\description{
Generate violin plot or box plot to show how errors are distributed by
proportion bins of 0.1. The errors can be displayed all mixed or splited by
cell type (\code{CellType}) or number of cell types present in the samples
(\code{nCellTypes}). See \code{facet.by} argument and examples for more
information.
}
\examples{
distErrorPlot(
  object = DDLSChungComp,
  error = "AbsErr",
  facet.by = "CellType",
  color.by = "nCellTypes",
  error.label = TRUE
)

distErrorPlot(
  object = DDLSChungComp,
  error = "AbsErr",
  x.by = "CellType",
  facet.by = NULL,
  filter.sc = FALSE,
  color.by = "CellType",
  error.label = TRUE
)
}
\seealso{
\code{\link{calculateEvalMetrics}} \code{\link{corrExpPredPlot}}
\code{\link{blandAltmanLehPlot}} \code{\link{barErrorPlot}}
}
