% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{exclusive.types}
\alias{exclusive.types}
\alias{exclusive.types<-}
\title{Get and set \code{exclusive.types} slot in a
\code{\linkS4class{ProbMatrixCellTypes}} object.}
\usage{
exclusive.types(object)

exclusive.types(object) <- value
}
\arguments{
\item{object}{A \code{\linkS4class{ProbMatrixCellTypes}} object.}

\item{value}{Optional slot that contains the exclusive cell types on the
experiment if they are provided. NULL by default.}
}
\description{
Get and set \code{exclusive.types} slot in a
\code{\linkS4class{ProbMatrixCellTypes}} object.
}
