% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{features}
\alias{features}
\alias{features<-}
\title{Get and set \code{features} slot in a
\code{\linkS4class{DigitalDLSorterDNN}} object}
\usage{
features(object)

features(object) <- value
}
\arguments{
\item{object}{A \code{\linkS4class{DigitalDLSorterDNN}} object.}

\item{value}{A \code{vector} with features (genes) considered by Deep Neural
Network model.}
}
\description{
Get and set \code{features} slot in a
\code{\linkS4class{DigitalDLSorterDNN}} object
}
