% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBulk.R
\name{generateBulkCellMatrix}
\alias{generateBulkCellMatrix}
\title{Generate training and test cell composition matrices}
\usage{
generateBulkCellMatrix(
  object,
  cell.ID.column,
  cell.type.column,
  prob.design,
  num.bulk.samples,
  n.cells = 100,
  train.freq.cells = 2/3,
  train.freq.bulk = 2/3,
  proportions.train = c(10, 5, 20, 15, 35, 15),
  proportions.test = c(10, 5, 20, 15, 35, 15),
  prob.zero = c(0.5, 0.5, 0.5, 0.5, 0.5, 0.5),
  balanced.type.cells = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object with
\code{single.cell.real} slot and, optionally, with \code{single.cell.simul}
slot.}

\item{cell.ID.column}{Name or column number corresponding to cell names of
expression matrix in cells metadata.}

\item{cell.type.column}{Name or column number corresponding to cell type of
each cell in cells metadata.}

\item{prob.design}{data.frame with the frequency ranges expected for each
cell type present in the experiment. This information can be estimated from
literature or from the single-cell experiment itself. This data.frame must
be built by three columns with specific headers (see examples): \itemize{
\item A cell type column with the same name of the cell type column in
cells metadata (\code{cell.type.column}). If the name of the column is not
the same, function returns an error. All cell types must appear in cells
metadata. \item A second column named \code{'from'} with the start
frequency for each cell type. \item A third column named \code{'to'} with
the final frequency for each cell type.}}

\item{num.bulk.samples}{Number of bulk RNA-seq samples proportions (and
therefore, simulated bulk RNA-seq samples) that will be generated taking
into account training and test data. We recommend seting this value
according to the number of single-cell profiles available in
\code{\linkS4class{DigitalDLSorter}} object avoiding an excesive
re-sampling, but generating a big number of samples for a better training.}

\item{n.cells}{Number of cells that will be aggregated in order to simulate
one bulk RNA-seq sample (100 by default).}

\item{train.freq.cells}{Proportion of cells used for simulating training bulk
samples (2/3 by default).}

\item{train.freq.bulk}{Proportion of bulk RNA-seq samples from the total
number (\code{num.bulk.samples}) used for training set (2/3 by default).}

\item{proportions.train}{Vector of five integers that determines the
proportions of bulk samples generated by the different methods (see Details
and Torroja and Sánchez-Cabo, 2019 for more information). This vector
represents proportions, so its entries must add up 100. By default, a
majority of random samples without using predefined ranges will be
generated.}

\item{proportions.test}{\code{proportions.train} for test samples.}

\item{prob.zero}{Probability of producing cell type proportions equal to
zero. It is a vector of five elements corresponding to the five methods for
production of cell type proportions (see \code{proportions.train} for more
details).}

\item{balanced.type.cells}{Boolean indicating if training and test cells will
be split in a balanced way considering cell types (\code{FALSE} by
default).}

\item{verbose}{Show informative messages during the execution (\code{TRUE} by
default).}
}
\value{
A \code{\linkS4class{DigitalDLSorter}} object with
\code{prob.cell.types} slot containing a \code{list} with two
\code{\linkS4class{ProbMatrixCellTypes}} objects (training and test). For
more information about the structure of this class, see
\code{?\linkS4class{ProbMatrixCellTypes}}.
}
\description{
Generate training and test cell composition matrices for the simulation of
bulk RNA-seq samples with known cell composition using single-cell expression
profiles. The resulting \code{\linkS4class{ProbMatrixCellTypes}} object
contains a matrix which determines the proportion of the different cell types
that will compose the simulated bulk samples. Moreover, this object also
contains other information relevant to the process. This function does not
simulate bulk samples, this task is performed by
\code{\link{simBulkProfiles}} or \code{\link{trainDigitalDLSorterModel}}
functions (see Documentation).
}
\details{
First, available single-cell profiles are split into training and test
subsets (2/3 for training and 1/3 for test by default (see
\code{train.freq.cells})) in order to avoid falsifying the results during the
evaluation of the model. Then, \code{num.bulk.samples} bulk samples
proportions are built and single-cell profiles that will be used for
simulating each bulk RNA-seq sample are set, being 100 cells per bulk sample
by default (see \code{n.cells} argument). The proportions of training and
test bulk samples are set by \code{train.freq.bulk} (2/3 for training and 1/3
for test by default). Finally, in order to avoid biases due to the
composition of the bulk RNA-seq, proportions of cell types
(\eqn{w_1,...,w_k}, where \eqn{k} is the number of cell types available in
single-cell profiles) are randomly generated by using five different
approaches:

\enumerate{ \item Cell proportions are randomly sampled from a truncated
uniform distribution with predefined limits according to a priori knowledge
of the abundance of each cell type (see \code{prob.design} argument). This
information can be inferred from the single-cell experiment itself or from
the literature. \item A second set is generated by randomly permuting cell
type labels from a distribution generated by the previous method. \item Cell
proportions are randomly sampled as by method 1 without replacement. \item
Using the last method for generating proportions, cell types labels are
randomly sampled. \item Cell proportions are randomly sampled from a
Dirichlet distribution.}

If you want to inspect the distribution of cell type proportions generated by
each method during the process, they can be shown by
\code{\link{showProbPlot}} function (see Documentation).
}
\examples{
\dontrun{
# generate a data.frame with frequency ranges of each cell type
probMatrix <- data.frame(
  Cell_type = c("ER+", "HER2+", "ER+ and HER2+", "TNBC",
                "Stromal", "Monocyte", "Tme", "BGC",
                "Bmem", "DC", "Macrophage", "TCD8", "Treg"),
  from = c(rep(30, 4), 1, rep(1, 8)),
  to = c(rep(70, 4), 50, rep(15, 8))
)
if (requireNamespace("digitalDLSorteRdata", quietly = TRUE)) {
  library(digitalDLSorteRdata)
  data(DDLSChung)
  DDLSChung <- generateBulkCellMatrix(
    object = DDLSChung,
    cell.type.column = "Cell_type",
    cell.ID.column = "Cell_ID",
    prob.design = probMatrix,
    num.bulk.samples = 100,
    verbose = TRUE
  )
  
}
}
}
\references{
Torroja, C. y Sánchez-Cabo, F. (2019). digitalDLSorter: A Deep
Learning algorithm to quantify immune cell populations based on scRNA-Seq
data. Frontiers in Genetics 10, 978. doi:
\href{https://doi.org/10.3389/fgene.2019.00978}{10.3389/fgene.2019.00978}
}
\seealso{
\code{\link{simBulkProfiles}}
\code{\linkS4class{ProbMatrixCellTypes}}
}
