% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getProbMatrix}
\alias{getProbMatrix}
\title{Getter function for the cell composition matrix}
\usage{
getProbMatrix(object, type.data)
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object with
\code{prob.cell.types} slot.}

\item{type.data}{Subset of data to e shown: \code{train} or \code{test}.}
}
\value{
A matrix object.
}
\description{
Getter function for the cell composition matrix. This function allows to
access to the cell composition matrix of simulated training or test
pseudo-bulk RNA-Seq data.
}
\seealso{
\code{\link{generateBulkCellMatrix}}
}
