% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadDeconvDataFromSummarizedExperiment}
\alias{loadDeconvDataFromSummarizedExperiment}
\title{Load data to deconvolute from a \code{\linkS4class{SummarizedExperiment}}
object}
\usage{
loadDeconvDataFromSummarizedExperiment(object, se.object, name.data = NULL)
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object with
\code{trained.model} slot.}

\item{se.object}{\code{\linkS4class{SummarizedExperiment}} object.}

\item{name.data}{Name with which the data is stored in
\code{\linkS4class{DigitalDLSorter}} object.}
}
\value{
\code{\linkS4class{DigitalDLSorter}} object with \code{deconv.data}
slot with the new bulk-RNAseq samples loaded
}
\description{
Load data in \code{\linkS4class{DigitalDLSorter}} object to deconvolute from
\code{\linkS4class{SummarizedExperiment}} object.
}
\seealso{
\code{\link{trainDigitalDLSorterModel}}
\code{\link{deconvDigitalDLSorterObj}}
}
