% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{model}
\alias{model}
\alias{model<-}
\title{Get and set \code{model} slot in a \code{DigitalDLSorterDNN} object.}
\usage{
model(object)

model(object) <- value
}
\arguments{
\item{object}{A \code{DigitalDLSorterDNN} object.}

\item{value}{A \code{keras.engine.sequential.Sequential} object with a
trained DNN model.}
}
\description{
Get and set \code{model} slot in a \code{DigitalDLSorterDNN} object.
}
