% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{model}
\alias{model}
\alias{model<-}
\title{Get and set \code{model} slot in a
\code{\linkS4class{DigitalDLSorterDNN}} object}
\usage{
model(object)

model(object) <- value
}
\arguments{
\item{object}{A \code{\linkS4class{DigitalDLSorterDNN}} object.}

\item{value}{A \code{keras.engine.sequential.Sequential} object with a
trained Deep Neural Network model.}
}
\description{
Get and set \code{model} slot in a
\code{\linkS4class{DigitalDLSorterDNN}} object
}
