% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{plots}
\alias{plots}
\alias{plots,ProbMatrixCellTypes-method}
\alias{plots<-}
\alias{plots<-,ProbMatrixCellTypes-method}
\title{Get and set \code{plots} slot in a
\code{\linkS4class{ProbMatrixCellTypes}} object}
\usage{
plots(object)

plots(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{ProbMatrixCellTypes}} object.}

\item{value}{List of lists with plots showing the distribution of cell
proportions generated by each method during the process.}
}
\description{
Get and set \code{plots} slot in a
\code{\linkS4class{ProbMatrixCellTypes}} object
}
