% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preparingToSave}
\alias{preparingToSave}
\title{Prepare \code{\linkS4class{DigitalDLSorter}} object to be saved as RDA file}
\usage{
preparingToSave(object)
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object with
\code{trained.data} slot.}
}
\description{
Prepare a \code{\linkS4class{DigitalDLSorter}} object which has a
\code{\linkS4class{DigitalDLSorterDNN}} object with trained DNN model.
\code{keras} models are not able to be stored natively as R objects (e.g.
RData or RDS files). By saving the structure as a JSON-like character object
and weights as a list, it is possible to recover the model and to carry out
predictions. \strong{Note:} With this option, the state of optimizer is not
saved, only architecture and weights.
}
\details{
It is possible to save completely the model as HDF5 file with
\code{\link{saveTrainedModelAsH5}} function and to load into
\code{\linkS4class{DigitalDLSorter}} object with
\code{\link{loadTrainedModelFromH5}} function.

It is also possible to save a \code{\linkS4class{DigitalDLSorter}} object as
RDS file with \code{saveRDS} function without any type of previous
preparation.
}
\seealso{
\code{\link{saveRDS}} \code{\link{saveTrainedModelAsH5}}
}
