% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{prob.cell.types}
\alias{prob.cell.types}
\alias{prob.cell.types<-}
\title{Get and set \code{prob.cell.types} slot in a \code{DigitalDLSorter}
object.}
\usage{
prob.cell.types(object, type.data = "both")

prob.cell.types(object, type.data = "both") <- value
}
\arguments{
\item{object}{A \code{DigitalDLSorter} object.}

\item{type.data}{Element of the list. Can be 'train', 'test' or 'both' (the
last by default).}

\item{value}{A list with two elements, train and test, each one with a
\code{ProbMatrixCellTypes} object.}
}
\description{
Get and set \code{prob.cell.types} slot in a \code{DigitalDLSorter}
object.
}
