% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ProbMatrixCellTypes-class}
\alias{ProbMatrixCellTypes-class}
\alias{ProbMatrixCellTypes}
\title{The Class ProbMatrixCellTypes}
\description{
The ProbMatrixCellTypes class is a data storage class that contains
information concerning to cell composition matrices used for the simulation
of bulk RNAseq samples. The matrix is saved in \code{prob.matrix} slot. The
rest of slots contain additional information generated during the process and
required in subsequent steps.
}
\details{
As described in Torroja and Sanchez-Cabo, 2019, the proportions are built
using five different methods in order to avoid biases due to the composition
of the simulated bulk samples. In \code{plots} slot, plots representing
visually the distribution of these probabilities are stored with the aim of
offering a way to monitor the different sets of samples generated. These
plots can be shown using \code{\link{showProbPlot}} function. See
\code{?\link{showProbPlot}} for more details.
}
\section{Slots}{

\describe{
\item{\code{prob.matrix}}{Matrix of cell proportions generated for the simulation of
bulk samples. Rows correspond to bulk samples which will be generated
(\eqn{i}), columns are the cell types present in the provided single-cell
data (\eqn{j}) and each entry is the proportion of \eqn{j} cell type in
\eqn{i} sample.}

\item{\code{cell.names}}{Matrix containing the names of the cells that will make up
each simulated bulk sample.}

\item{\code{set.list}}{List of cells ordered according to the cell type to which they
belong.}

\item{\code{set}}{Vector with cell names present in the object.}

\item{\code{plots}}{List of lists with plots showing the distribution of cell
proportions generated by each method during the process. In each list,
\code{boxplot}, \code{violinplot}, \code{linesplot} or \code{ncelltypes}
can be found. Please, see \code{\link{showProbPlot}} for more details.}

\item{\code{type.data}}{Character with the type of data contained: training or test.}
}}

\references{
Torroja, C. y Sánchez-Cabo, F. (2019). digitalDLSorter: A Deep
Learning algorithm to quantify immune cell populations based on scRNA-Seq
data. Frontiers in Genetics 10, 978. doi: \url{10.3389/fgene.2019.00978}
}
