% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{project}
\alias{project}
\alias{project<-}
\title{Get and set \code{project} slot in a \code{DigitalDLSorter}
object.}
\usage{
project(object)

project(object) <- value
}
\arguments{
\item{object}{A \code{DigitalDLSorter} object.}

\item{value}{A character indicating the name of the project.}
}
\description{
Get and set \code{project} slot in a \code{DigitalDLSorter}
object.
}
