% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{saveRDS}
\alias{saveRDS}
\title{Save \code{\linkS4class{DigitalDLSorter}} object as RDS file}
\usage{
saveRDS(
  object,
  file,
  ascii = FALSE,
  version = NULL,
  compress = TRUE,
  refhook = NULL
)
}
\description{
Save \code{\linkS4class{DigitalDLSorter}} and
\code{\linkS4class{DigitalDLSorterDNN}} objects as RDS files. We developed
this generic with the aim of changing the behavior of the base function and
saving the structure and weights of DNN model as R native objects. This is
because \pkg{keras} models are not able to be stored natively as R objects
(e.g. RData or RDS files). By saving the structure as JSON character object
and weights as list object, it is possible recovering the model and carrying
out predictions. If \code{trained.model} slot is empty, the function will
have the usual behavior.
}
\details{
With this option, the state of optimizer is not saved, only architecture and
weights. It is possible to save completely the model as HDF5 file with
\code{\link{saveTrainedModelAsH5}} function and to load into
\code{\linkS4class{DigitalDLSorter}} object with
\code{\link{loadTrainedModelFromH5}} function.

Moreover, if you want to save the object as RDA file, it is possible by
converting the model to an allowed R object with
\code{\link{preparingToSave}} function.
}
\seealso{
\code{\link{saveTrainedModelAsH5}} \code{\link{preparingToSave}}
}
