% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{saveRDS}
\alias{saveRDS}
\alias{saveRDS,saveRDS-method}
\alias{saveRDS,DigitalDLSorterDNN-method}
\alias{saveRDS,DigitalDLSorter-method}
\title{Save \code{\linkS4class{DigitalDLSorter}} object as RDS file}
\usage{
saveRDS(
  object,
  file,
  ascii = FALSE,
  version = NULL,
  compress = TRUE,
  refhook = NULL
)

\S4method{saveRDS}{DigitalDLSorterDNN}(
  object,
  file,
  ascii = FALSE,
  version = NULL,
  compress = TRUE,
  refhook = NULL
)

\S4method{saveRDS}{DigitalDLSorter}(
  object,
  file,
  ascii = FALSE,
  version = NULL,
  compress = TRUE,
  refhook = NULL
)
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} or
\code{\linkS4class{DigitalDLSorterDNN}} object to save}

\item{file}{File path where the object will be saved}

\item{ascii}{a logical.  If \code{TRUE} or \code{NA}, an ASCII
    representation is written; otherwise (default), a binary one is used.
    See the comments in the help for \code{\link[base]{save}}.}

\item{version}{the workspace format version to use.  \code{NULL}
    specifies the current default version (3). The only other supported
    value is 2, the default from \R 1.4.0 to \R 3.5.0.}

\item{compress}{a logical specifying whether saving to a named file is
    to use \code{"gzip"} compression, or one of \code{"gzip"},
    \code{"bzip2"} or \code{"xz"} to indicate the type of compression to
    be used.  Ignored if \code{file} is a connection.}

\item{refhook}{a hook function for handling reference objects.}
}
\description{
Save \code{\linkS4class{DigitalDLSorter}} and
\code{\linkS4class{DigitalDLSorterDNN}} objects as RDS files.
\pkg{keras} models are not able to be
stored natively as R objects (e.g. RData or RDS files). By saving the
structure as a JSON-like character object and weights as a list, it is
possible to recover the model and to perform predictions. If
\code{trained.model} slot is empty, the function will have the usual
behavior.
}
\details{
Note: with this option, the state of optimizer is not saved, only
architecture and weights. It is possible to save completely the model as HDF5
file with \code{\link{saveTrainedModelAsH5}} function and to load into a
\code{\linkS4class{DigitalDLSorter}} object with
\code{\link{loadTrainedModelFromH5}} function. See documentation for details.
}
\seealso{
\code{\linkS4class{DigitalDLSorter}}
\code{\link{saveTrainedModelAsH5}}
}
