% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnnModel.R
\name{saveTrainedModelAsH5}
\alias{saveTrainedModelAsH5}
\title{Save on disk trained DigitalDLSorter DNN model as HDF5 file.}
\usage{
saveTrainedModelAsH5(object, file.path, overwrite = FALSE)
}
\arguments{
\item{object}{\code{\link{DigitalDLSorter}} object with \code{trained.model}
slot.}

\item{file.path}{Valid file path where saving the model.}

\item{overwrite}{Overwrite file if it already exists.}
}
\description{
Save on disk the trained model in HDF5 format. Note that this function does
not save the \code{\link{DigitalDLSorterDNN}} object, but the trained keras
model.
}
\seealso{
\code{\link{trainDigitalDLSorterModel}}
\code{\link{loadTrainedModelFromH5}}
}
