% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{set}
\alias{set}
\alias{set,ProbMatrixCellTypes-method}
\alias{set<-}
\alias{set<-,ProbMatrixCellTypes-method}
\title{Get and set \code{set} slot in a
\code{\linkS4class{ProbMatrixCellTypes}} object}
\usage{
set(object)

set(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{ProbMatrixCellTypes}} object.}

\item{value}{Vector with names of cells present in the object.}
}
\description{
Get and set \code{set} slot in a
\code{\linkS4class{ProbMatrixCellTypes}} object
}
