% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showProbPlot}
\alias{showProbPlot}
\title{Show distribution plots of the cell proportions generated by
\code{\link{generateBulkCellMatrix}}}
\usage{
showProbPlot(object, type.data, set, type.plot = "boxplot")
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object with
\code{prob.cell.types} slot with \code{plot} slot.}

\item{type.data}{Subset of data to show: \code{train} or \code{test}.}

\item{set}{Integer determining which of the 6 different subsets to display.}

\item{type.plot}{Character determining which type of visualization to
display. It can be \code{'boxplot'}, \code{'violinplot'}, \code{'linesplot'} or
\code{'ncelltypes'}. See Description for more information.}
}
\value{
\code{ggplot} object.
}
\description{
Show distribution plots of the cell proportions generated by
\code{\link{generateBulkCellMatrix}}. These frequencies will determine the
proportion of different cell types used during the simulation of pseudo-bulk
RNA-Seq samples. There are 6 subsets of proportions generated by different
approaches that can be visualized in three ways: box plots, violin plots and
lines plots. You can also plot the probabilities based on the number of
different cell types present in the samples by setting \code{type.plot =
'nCellTypes'}.
}
\details{
These plots are only for diagnostic purposes. This is the reason because they
are generated without any parameter introduced by the user.
}
\examples{
if (requireNamespace("digitalDLSorteRdata", quietly = TRUE)) {
  library(digitalDLSorteRdata)
  data(DDLSLi.list)
  DDLSLi <- listToDDLS(DDLSLi.list)
  lapply(
    X = 1:6, FUN = function(x) {
      showProbPlot(
        DDLSLi,
        type.data = "train",
        set = x,
        type.plot = "boxplot"
      )
    }
  )
}
}
\seealso{
\code{\link{generateBulkCellMatrix}}
}
