% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showProbPlot}
\alias{showProbPlot}
\title{Show distribution plots of cell proportions generated by
\code{\link{generateBulkCellMatrix}}}
\usage{
showProbPlot(object, type.data, set, type.plot = "boxplot")
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object with
\code{prob.cell.types} slot with \code{plot} slot.}

\item{type.data}{Subset of data to show: \code{train} or \code{test}.}

\item{set}{Integer determining which of the 6 different subsets is to be
shown.}

\item{type.plot}{Character determining which type of visualization is to be
shown. It can be \code{boxplot}, \code{violinplot}, \code{linesplot} or
\code{ncelltypes}. See Description for more information.}
}
\value{
\code{ggplot} object.
}
\description{
Show distribution plots of cell proportions generated by
\code{\link{generateBulkCellMatrix}}. These frequencies will determine the
proportion of different cell types used during the simulation of bulk RNA-Seq
samples. There are 6 subsets of proportions (1 and 2 are generated by the
same way) generated by different approaches that can be visualized in three
ways: boxplot, violinplot and linesplot. You also can represent the
probabilities based on the number of different cell types present in samples
setting \code{type.plot = 'nCellTypes'}.
}
\details{
These plots are only for diagnostic purposes. This is the reason because they
are generated without any parameter introduced by the user.
}
\examples{
lapply(
  X = 1:6, FUN = function(x) {
    showProbPlot(
      DDLSChungComp,
      type.data = "train",
      set = x,
      type.plot = "boxplot"
    )
  }
)

}
\seealso{
\code{\link{generateBulkCellMatrix}}
}
