% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSingleCell.R
\name{simSingleCellProfiles}
\alias{simSingleCellProfiles}
\title{Simulate new single-cell expression profiles using the estimated ZINB
parameters.}
\usage{
simSingleCellProfiles(
  object,
  cell.ID.column,
  cell.type.column,
  n.cells,
  verbose = TRUE
)
}
\arguments{
\item{object}{\code{\link{DigitalDLSorter}} object with
\code{single.cell.real} and \code{zinb.params} slots.}

\item{cell.ID.column}{Name or number of the column in cells metadata
corresponding with cell names in expression matrix.}

\item{cell.type.column}{Name or number of the column in cells metadata
corresponding with the cell type of each cell.}

\item{n.cells}{Number of simulated cells generated by cell type (i.e. if you
have 10 different cell types on your dataset, if \code{n.cells = 100}, then
1000 cell profiles will be simulated).}

\item{verbose}{Show informative messages during the execution.}
}
\value{
A \code{\link{DigitalDLSorter}} object with \code{single.cell.final}
slot containing a \code{SingleCellExperiment} object with the simulated
single-cell profiles.
}
\description{
Simulate single-cell expression profiles by randomly sampling from a negative
binomial distribution using ZINB parameters estimated by ZINB-WaVE model and
introducing dropouts by sampling from a binomial distribution with ZINB-WaVE
model estimated.
}
\details{
Before this step, see \code{\link{estimateZinbwaveParams}}. As described in
Torroja and Sanchez-Cabo, 2019, this function simulates a determined number
of transcriptional profiles for each cell type provided by randomly sampling
from a negative binomial distribution with \eqn{\mu} and \eqn{\theta}
estimated parameters and introducing dropouts by sampling from a binomial
distribution with pi probability. All paramteres are estimated from
single-cell real data using \code{\link{estimateZinbwaveParams}} function. It
uses the ZINB-WaVE model (Risso et al., 2018). For more details about the
model, see \code{\link{estimateZinbwaveParams}}.
}
\examples{
DDLSSmallCompleted <- simSingleCellProfiles(
  object = DDLSSmallCompleted,
  cell.ID.column = "Cell_ID",
  cell.type.column = "Cell_type",
  n.cells = 10,
  verbose = TRUE
)

}
\references{
Risso, D., Perraudeau, F., Gribkova, S. et al. (2018). A general
and flexible method for signal extraction from single-cell RNA-seq data.
Nat Commun 9, 284. doi: \url{10.1038/s41467-017-02554-5}.

Torroja, C. y Sánchez-Cabo, F. (2019). digitalDLSorter: A Deep Learning
algorithm to quantify immune cell populations based on scRNA-Seq data.
Frontiers in Genetics 10, 978. doi: \url{10.3389/fgene.2019.00978}
}
\seealso{
\code{\link{estimateZinbwaveParams}}
}
