% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{single.cell.simul}
\alias{single.cell.simul}
\alias{single.cell.simul,DigitalDLSorter-method}
\alias{single.cell.simul<-}
\alias{single.cell.simul<-,DigitalDLSorter-method}
\title{Get and set \code{single.cell.simul} slot in a
\code{\linkS4class{DigitalDLSorter}} object}
\usage{
single.cell.simul(object)

single.cell.simul(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object.}

\item{value}{\code{\linkS4class{SingleCellExperiment}} object with simulated
single-cell profiles.}
}
\description{
Get and set \code{single.cell.simul} slot in a
\code{\linkS4class{DigitalDLSorter}} object
}
