% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{test.metrics}
\alias{test.metrics}
\alias{test.metrics<-}
\title{Get and set \code{test.metrics} slot in a
\code{DigitalDLSorterDNN} object.}
\usage{
test.metrics(object)

test.metrics(object) <- value
}
\arguments{
\item{object}{A \code{DigitalDLSorterDNN} object.}

\item{value}{A \code{list} object with the resulting metrics after prediction
on test data with DNN model.}
}
\description{
Get and set \code{test.metrics} slot in a
\code{DigitalDLSorterDNN} object.
}
