% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{test.metrics}
\alias{test.metrics}
\alias{test.metrics,DigitalDLSorterDNN-method}
\alias{test.metrics<-}
\alias{test.metrics<-,DigitalDLSorterDNN-method}
\title{Get and set \code{test.metrics} slot in a
\code{\linkS4class{DigitalDLSorterDNN}} object}
\usage{
test.metrics(object)

test.metrics(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorterDNN}} object.}

\item{value}{List object with the resulting metrics after prediction
on test data with the Deep Neural Network model.}
}
\description{
Get and set \code{test.metrics} slot in a
\code{\linkS4class{DigitalDLSorterDNN}} object
}
