% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{trained.model}
\alias{trained.model}
\alias{trained.model<-}
\title{Get and set \code{trained.model} slot in a
\code{\linkS4class{DigitalDLSorter}} object}
\usage{
trained.model(object)

trained.model(object) <- value
}
\arguments{
\item{object}{A \code{\linkS4class{DigitalDLSorter}} object.}

\item{value}{A \code{\linkS4class{DigitalDLSorterDNN}} object.}
}
\description{
Get and set \code{trained.model} slot in a
\code{\linkS4class{DigitalDLSorter}} object
}
