% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{trained.model}
\alias{trained.model}
\alias{trained.model,DigitalDLSorter-method}
\alias{trained.model<-}
\alias{trained.model<-,DigitalDLSorter-method}
\title{Get and set \code{trained.model} slot in a
\code{\linkS4class{DigitalDLSorter}} object}
\usage{
trained.model(object)

trained.model(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{DigitalDLSorter}} object.}

\item{value}{\code{\linkS4class{DigitalDLSorter}} object.}
}
\description{
Get and set \code{trained.model} slot in a
\code{\linkS4class{DigitalDLSorter}} object
}
